% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBC.R
\docType{data}
\name{ACBC}
\alias{ACBC}
\title{ACBC example dataset}
\format{
A data frame with 95 rows and 39 variables
}
\usage{
data(ACBC)
}
\description{
A data set with raw scores export for an adaptive choice-based conjoint.
}
\details{
\itemize{
  \item ID. \code{ID} of participants
  \item RLH. Root likelihood.
  \item NumParameters. Number of parameters per respondent
  \item Att1_Lev1. part-worth utility of Level 1 of Attribute 1
  \item Att1_Lev2. part-worth utility of Level 2 of Attribute 1
  \item Att1_Lev3. part-worth utility of Level 3 of Attribute 1
  \item Att1_Lev4. part-worth utility of Level 4 of Attribute 1
  \item Att1_Lev5. part-worth utility of Level 5 of Attribute 1
  \item Att2_Lev1. part-worth utility of Level 1 of Attribute 2
  \item Att2_Lev2. part-worth utility of Level 2 of Attribute 2
  \item Att2_Lev3. part-worth utility of Level 3 of Attribute 2
  \item Att2_Lev4. part-worth utility of Level 4 of Attribute 2
  \item Att3_Lev1. part-worth utility of Level 1 of Attribute 3
  \item Att3_Lev2. part-worth utility of Level 2 of Attribute 3
  \item Att3_Lev3. part-worth utility of Level 3 of Attribute 3
  \item Att3_Lev4. part-worth utility of Level 4 of Attribute 3
  \item Att4_Lev1. part-worth utility of Level 1 of Attribute 4
  \item Att4_Lev2. part-worth utility of Level 2 of Attribute 4
  \item Att4_Lev3. part-worth utility of Level 3 of Attribute 4
  \item Att5_Lev1. part-worth utility of Level 1 of Attribute 5
  \item Att5_Lev2. part-worth utility of Level 2 of Attribute 5
  \item Att5_Lev3. part-worth utility of Level 3 of Attribute 5
  \item Att5_Lev4. part-worth utility of Level 4 of Attribute 5
  \item Att6_Lev1. part-worth utility of Level 1 of Attribute 6
  \item Att6_Lev2. part-worth utility of Level 2 of Attribute 6
  \item Att6_Lev3. part-worth utility of Level 3 of Attribute 6
  \item Att6_Lev4. part-worth utility of Level 4 of Attribute 6
  \item Att7_Lev1. part-worth utility of Level 1 of Attribute 7
  \item Att7_Lev2. part-worth utility of Level 2 of Attribute 7
  \item Att7_Lev3. part-worth utility of Level 3 of Attribute 7
  \item Att7_Lev4. part-worth utility of Level 4 of Attribute 7
  \item Att7_Lev5. part-worth utility of Level 5 of Attribute 7
  \item Att8_Lev1. part-worth utility of Level 1 of Attribute 8
  \item Att8_Lev2. part-worth utility of Level 2 of Attribute 8
  \item price_2.093. linear coding for price at $2.093 (lower bound)
  \item price_27.287. linear coding for price at $27.287 (upper bound)
  \item none. part-worth utility of none parameter
  \item HOT. actual choice in the Holdout task
  \item Group. Grouping variable
}
}
\keyword{datasets}
