% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACBC_prep.R
\docType{data}
\name{ACBC_prep}
\alias{ACBC_prep}
\title{ACBC example dataset}
\format{
A data frame with 80 rows and 34 variables
}
\usage{
data(ACBC_prep)
}
\description{
A data set with raw scores export for an ACBC.
}
\details{
\itemize{
  \item ID. \code{ID} of participants
  \item RLH. Root likelihood.
  \item NumParameters. Number of parameters per respondent
  \item Att1_Lev1. part-worth utility of Level 1 of Attribute 1
  \item Att1_Lev2. part-worth utility of Level 2 of Attribute 1
  \item Att1_Lev3. part-worth utility of Level 3 of Attribute 1
  \item Att1_Lev4. part-worth utility of Level 4 of Attribute 1
  \item Att2_Lev1. part-worth utility of Level 1 of Attribute 2
  \item Att2_Lev2. part-worth utility of Level 2 of Attribute 2
  \item Att2_Lev3. part-worth utility of Level 3 of Attribute 2
  \item Att2_Lev4. part-worth utility of Level 4 of Attribute 2
  \item Att2_Lev5. part-worth utility of Level 5 of Attribute 2
  \item Att3_Lev1. part-worth utility of Level 1 of Attribute 3
  \item Att3_Lev2. part-worth utility of Level 2 of Attribute 3
  \item Att3_Lev3. part-worth utility of Level 3 of Attribute 3
  \item Att3_Lev4. part-worth utility of Level 4 of Attribute 3
  \item Att4_Lev1. part-worth utility of Level 1 of Attribute 4
  \item Att4_Lev2. part-worth utility of Level 2 of Attribute 4
  \item Att4_Lev3. part-worth utility of Level 3 of Attribute 4
  \item Att4_Lev4. part-worth utility of Level 4 of Attribute 4
  \item Att4_Lev5. part-worth utility of Level 5 of Attribute 4
  \item Att4_Lev6. part-worth utility of Level 6 of Attribute 4
  \item Att4_Lev7. part-worth utility of Level 7 of Attribute 4
  \item Att5_Lev1. part-worth utility of Level 1 of Attribute 5
  \item Att5_Lev2. part-worth utility of Level 2 of Attribute 5
  \item Att5_Lev3. part-worth utility of Level 3 of Attribute 5
  \item Att5_Lev4. part-worth utility of Level 4 of Attribute 5
  \item Att6_Lev1. part-worth utility of Level 1 of Attribute 6
  \item Att6_Lev2. part-worth utility of Level 2 of Attribute 6
  \item price_low. part-worth utilitiy of lower price level
  \item price_high. part-worth utilitiy of upper price level
  \item none. part-worth utilitiy of none parameter
  \item HOT. actual choice in the Holdout task
  \item Group. Grouping variable
}
}
\keyword{datasets}
