% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{Accuracy}
\alias{Accuracy}
\title{Accuracy}
\usage{
Accuracy(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame including Holdout Options and actual \code{"choice"} and \code{"Group"} if optional argument is defined}

\item{id}{column index}

\item{Group}{optional grouping variable to get accuracy by group}

\item{opts}{column indexes of the options included in the holdout task}

\item{choice}{column index of the actual choice}

\item{None}{column index}
}
\value{
test
}
\description{
Accuracy is defined as Number of correct predicted participants divided by the Total number of predictions
}
