% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CBC.R
\docType{data}
\name{CBC}
\alias{CBC}
\title{CBC example dataset}
\format{
A data frame with 79 rows and 23 variables
}
\usage{
data(CBC)
}
\description{
A data set with raw scores export for an CBC, assuming part-worth coding for
price (if included).
}
\details{
\itemize{
  \item ID. \code{ID} of participants
  \item RLH. Root likelihood.
  \item NumParameters. Number of parameters per respondent
  \item Att1_Lev1. part-worth utility of Level 1 of Attribute 1
  \item Att1_Lev2. part-worth utility of Level 2 of Attribute 1
  \item Att1_Lev3. part-worth utility of Level 3 of Attribute 1
  \item Att1_Lev4. part-worth utility of Level 4 of Attribute 1
  \item Att1_Lev5. part-worth utility of Level 5 of Attribute 1
  \item Att2_Lev1. part-worth utility of Level 1 of Attribute 2
  \item Att2_Lev2. part-worth utility of Level 2 of Attribute 2
  \item Att2_Lev3. part-worth utility of Level 3 of Attribute 2
  \item Att2_Lev4. part-worth utility of Level 4 of Attribute 2
  \item Att2_Lev5. part-worth utility of Level 5 of Attribute 2
  \item Att2_Lev6. part-worth utility of Level 6 of Attribute 2
  \item Att3_Lev1. part-worth utility of Level 1 of Attribute 3
  \item Att3_Lev2. part-worth utility of Level 2 of Attribute 3
  \item Att3_Lev3. part-worth utility of Level 3 of Attribute 3
  \item Att3_Lev4. part-worth utility of Level 4 of Attribute 3
  \item Att3_Lev5. part-worth utility of Level 5 of Attribute 3
  \item Att3_Lev6. part-worth utility of Level 6 of Attribute 3
  \item Att3_Lev7. part-worth utility of Level 7 of Attribute 3
  \item none. part-worth utility of none parameter
  \item HOT. actual choice in the Holdout task
  \item Group. Grouping variable
}
}
\keyword{datasets}
