% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHOT.R
\name{createHOT}
\alias{createHOT}
\title{Preparing Holdout task and creating utilities}
\usage{
createHOT(
  data,
  id,
  None = NULL,
  prod,
  prod.levels,
  interpolate.levels = NULL,
  piece.p = NULL,
  lin.p = NULL,
  coding = NULL,
  method = c("ACBC" | "CBC" | "MaxDiff"),
  varskeep = NULL,
  choice
)
}
\arguments{
\item{data}{a data frame}

\item{id}{the column index of \code{id} in \code{data}}

\item{None}{the column index of \code{None} in \code{data}; if of \code{None} is not included, leave empty}

\item{prod}{number of options in the Holdout task without the \code{None} option, must be numeric}

\item{prod.levels}{a list to define the attribute levels of the options (\code{prod})}

\item{interpolate.levels}{a list of the levels of the variables that should be interpolated. These needs to be the same as provided to Sawtooth Software. Please make sure to provide the whole list. Only needs to be specified for the variables that are coded as 1 (linear) or 2 (piecewise)}

\item{piece.p}{a list of the column indexes of the lower level and the upper level that should be used for interpolating}

\item{lin.p}{vector of the column indexes of the linear variables}

\item{coding}{vector of the coding of each attribute, 0 = part-worth coding, 1 = linear coding, 2 = piecewise coding; please make sure to code linear price of ACBC as piecewise since you have two values to interpolate}

\item{method}{specify the \code{method} your study; needs to be one of the following: MaxDiff, CBC, or ACBC}

\item{varskeep}{variables that should be kept in the data frame, use column index}

\item{choice}{actual choice in the Holdout task}
}
\value{
a data frame
}
\description{
Function used to create utilities for validation task.
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(
  data = MaxDiff, None = 19,
  id = 1, prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  choice = 20, method = "MaxDiff"
)

}
