% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createUtils.R
\name{createUtil}
\alias{createUtil}
\title{Preparing Holdout task and creating utilities}
\usage{
createUtil(
  data,
  id,
  None = NULL,
  prod,
  x,
  method = c("ACBC" | "CBC" | "MaxDiff"),
  price = NULL,
  price_low = NULL,
  price_high = NULL,
  price_low_po = NULL,
  price_high_po = NULL,
  varskeep = NULL
)
}
\arguments{
\item{data}{a data frame}

\item{id}{the column index of \code{id} in \code{data}}

\item{None}{the column index of \code{None} in \code{data}; if of \code{None} is not included, leave empty}

\item{prod}{number of options in the Holdout task without the \code{None} option, must be numeric}

\item{x}{define the attribute levels of the products}

\item{method}{specify the \code{method} your study; needs to be one of the following: MaxDiff, CBC, or ACBC}

\item{price}{whether you use \code{"fixed"} price level or \code{"interpolate"} if you use \code{"method = 'ACBC'"}, if you use \code{"method = 'CBC'"} please set \code{"price"}
either to \code{"linear"} or \code{"piecewise"}}

\item{price_low}{lower price border}

\item{price_high}{upper price border}

\item{price_low_po}{column index of lower price}

\item{price_high_po}{column index of upper price}

\item{varskeep}{variables that should be kept in the data frame, use column index}
}
\value{
a data frame
}
\description{
Function used to create utilities for a specific Holdout task.
}
