% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freq_assort.R
\name{freqassort}
\alias{freqassort}
\title{Function to calculate averaged number of times a participant is reached by
assortment}
\usage{
freqassort(data, group, none, opts)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{freqassort()} by group(s).}

\item{none}{Column name of the none / threshold alternative.}

\item{opts}{Column names of the alternatives included in the assortment.}
}
\value{
a tibble
}
\description{
Frequency function of T(otal) U(nduplicated) R(each) and F(requency)
analysis is designed to measure the average time a consumer is reached
with a particular product bundle assortment.
}
\details{
\code{freqassort()} calculates the average times a consumer would be reached with
the tested product assortment. The current logic of \code{freqassort()} is that
the utility of an alternative has to exceed a threshold.
In the case of \code{freqassort()}, this threshold is referred
to the \code{none} argument in \code{data}. \code{none} does not necessarily have to be a
no-buy but it could also be, e.g., a current product.

\code{data} a \code{data.frame} object containing the alternatives that should
be tested.

\code{group} optional grouping variable(s) to display results by group(s).
Has to be the column name(s) of variables in \code{data}.

\code{opts} defines product assortment that should be considered.
Input of \code{opts} has to be column names of variables in \code{data}.

\code{none} to specify the column name of the \code{none} alternative (i.e.,
threshold variable).
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  varskeep = "group",
  choice = "hot"
)

# freqassort - without group argument defined
freqassort(
  data = hot,
  opts = c(option_1, option_2, option_6),
  none = none
)

# freqassort - with group argument defined
freqassort(
  data = hot,
  opts = c(option_1, option_2, option_6),
  none = none,
  group = group
)

}
