% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{kl}
\alias{kl}
\title{Function to calculate Kullback-Leibler divergence of validation/holdout
task}
\usage{
kl(data, group, opts, choice, epsilon = NULL, base = NULL)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get `kl()` by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice in the validation/holdout
task.}

\item{epsilon}{A vector of noise that should be added to 0 values, per
default set to 1e-05 (see Drost, 2018).}

\item{base}{A character string to define the logarithm base, currently two
are provided, namely `log` (default) and `log2`.}
}
\value{
a tibble
}
\description{
Function to measure the Kullback-Leibler Divergence of a
validation/holdout task.
}
\details{
Kullback-Leibler-Divergence measures the divergence between the actual
choice distribution and the predicted choice distribution (Ding et al., 2011;
Drost, 2018). Currently only provides the deviation measured based on
\eqn{log} and \eqn{log{_2}} algorithm. \eqn{log} set as default.

Due to Kullback-Leibler divergence's asymmetry, the output provides both
`KL_O_P` which is equivalent to (Observed || Predicted) and
`KL_P_O` which is equivalent to (Predicted || Observed).

`data` has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the `createHOT()`
function.

`group` optional grouping variable, if results should be displayed
by different groups. Has to be column name of variables in `data`.

`opts` is required to specify the different alternatives in the
validation/holdout task (also includes the `none` alternative).
Input of `opts` has to be column names of variables in `data`.

`choice` to specify column of actual choice in the validation/holdout
task. Input of opts `choice` has to be column name of actual choice.

`epsilon` has to be a numeric input in case of 0 in the numerator or
denominator. 0 then will be replaced by `epsilon`. Default value
is `epsilon = 1e-5`, however, can be adopted (see also Drost, 2018).

`base` has to be a character string, deciding which logarithm base
you want to apply to calculate Kullback-Leibler. You can choose
between \eqn{log} and \eqn{log{_2}}. Default is set to \eqn{log}.
}
\examples{

library(validateHOT)

HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# kl with log base - without group argument defined
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log"
)

\dontrun{
# kl with log2 base - without group argument defined
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log2"
)

# kl grouped - log + specifying epsilon
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log",
  group = Group,
  epsilon = 1e-8
)

# kl grouped - log2
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log2",
  group = Group
)
}

}
\references{
{

Ding, Min, John R. Hauser, Songting Dong, Daria Dzyabura, Zhilin Yang,
SU Chenting, and Steven P. Gaskin. (2011).
Unstructured Direct Elicitation of Decision Rules. \emph{Journal of
Marketing Research 48}(1): 116-27.
\verb{https://doi.org/10.1509/jmkr.48.1.116}.


Drost, Hajk-Georg. (2018). Philentropy: Information Theory and Distance
Quantification with R. \emph{Journal of Open Source Software 3}(26),
765, \verb{https://joss.theoj.org/papers/10.21105/joss.00765}.

}
}
