% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MaxDiff.R
\docType{data}
\name{MaxDiff}
\alias{MaxDiff}
\title{MaxDiff example dataset}
\format{
A data frame with 70 rows and 21 variables
}
\usage{
data(MaxDiff)
}
\description{
A data set with raw scores export for a MaxDiff.
}
\details{
\itemize{
  \item ID. \code{ID} of participants
  \item RLH. Root likelihood.
  \item Option_01. part-worth utility of MaxDiff object # 1
  \item Option_02. part-worth utility of MaxDiff object # 2
  \item Option_03. part-worth utility of MaxDiff object # 3
  \item Option_04. part-worth utility of MaxDiff object # 4
  \item Option_05. part-worth utility of MaxDiff object # 5
  \item Option_06. part-worth utility of MaxDiff object # 6
  \item Option_07. part-worth utility of MaxDiff object # 7
  \item Option_08. part-worth utility of MaxDiff object # 8
  \item Option_09. part-worth utility of MaxDiff object # 9
  \item Option_10. part-worth utility of MaxDiff object # 10
  \item Option_11. part-worth utility of MaxDiff object # 11
  \item Option_12. part-worth utility of MaxDiff object # 12
  \item Option_13. part-worth utility of MaxDiff object # 13
  \item Option_14. part-worth utility of MaxDiff object # 14
  \item Option_15. part-worth utility of MaxDiff object # 15
  \item Option_16. part-worth utility of MaxDiff object # 16
  \item none. part-worth utilitiy of none parameter
  \item HOT. actual choice in the Holdout task
  \item Group. Grouping variable
}
}
\keyword{datasets}
