% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedAE.R
\name{medae}
\alias{medae}
\title{Function to calculate median absolute error}
\usage{
medae(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get `medae()` by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice in the validation/holdout
task.}
}
\value{
a tibble
}
\description{
`medae` measures the median absolute error (MedAE) of a
validation/holdout task.
}
\details{
`medae` calculates the deviation between predicted and
stated (actual) choice share and returns the median error in the
validation/holdout task, which is less likely to be influenced by outliers.

`data` has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the `createHOT()`
function.

`group` optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in `data`.

`opts` is required to specify the different alternatives in the
validation/holdout task. Input of `opts` has to be column names
of variables in `data`.

`choice` to specify column of actual choice in the validation/holdout
task. Input of opts `choice` has to be column name of actual choice.
}
\examples{

library(validateHOT)

HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# medae - without group argument defined
medae(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice
)

# medae - with group argument defined
medae(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  group = Group
)

}
