% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhp.R
\name{mhp}
\alias{mhp}
\title{Function to calculate Mean Hit Probability (MHP) of validation task}
\usage{
mhp(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to calculate \code{mhp()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}
}
\value{
a tibble
}
\description{
\code{mhp()} calculates the mean hit probability of a validation
task.
}
\details{
Mean hit probability (MHP) calculates the averaged hit probability of
participants actual choices in the validation task.

\code{data} a \code{data.frame} object including the alternatives shown in the
validation task. It can be created using the \code{create_hot()} function.

\code{group} optional grouping variable(s) to display results by group(s).
different groups. Has to be the column name(s) of variables in \code{data}.

\code{opts} to specify the different alternatives in the
validation task (also includes the \code{none} alternative).

\code{choice} to specify the column of actual choice in the validation
task. The input of \code{choice} has to be the column name of actual choice.

Output will display both mean hit probability and its corresponding standard
error (both in percentage).
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  choice = "hot",
  varskeep = "group"
)

# mhp - without group argument defined
mhp(
  data = hot,
  opts = c(option_1:none),
  choice = choice
)

# mhp - with group argument defined
mhp(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  group = group
)

}
