% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precision.R
\name{precision}
\alias{precision}
\title{Function to calculate precision of confusion matrix for general product
demand}
\usage{
precision(data, group, opts, choice, none)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{precision()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}

\item{none}{Column name of none alternative.}
}
\value{
a tibble
}
\description{
\code{precision()} is one of the 5 metrics of the confusion matrix
and is defined as \eqn{\frac{TP}{TP + FP}}, where TP =
True Positives, FP = False Positives (see, e.g., Burger, 2018).
}
\details{
The current logic of \code{precision()} is to determine whether a binary-coded
choice is correctly predicted by the model. To use the function, the
validation task must include a \code{none} alternative.
A potential application could involve evaluating whether a buy or a no-buy
choice has been correctly predicted. For example, consider a scenario with
three alternatives plus a \code{none} alternative. You may want to verify if the
model correctly predicted a buy or no-buy decision (i.e., overestimates or
underestimates the purchase likelihood). The \code{precision()} function can be
useful for assessing the precision of that 2x2 matrix.

\code{data} a \code{data.frame} object including the alternatives shown in the
validation task. It can be created using the \code{create_hot()} function.

\code{group} optional grouping variable(s) to display results by group(s).
different groups. Has to be the column name(s) of variables in \code{data}.

\code{opts} to specify the different alternatives in the
validation task (also includes the \code{none} alternative).
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify the column of actual choice in the validation
task. The input of \code{choice} has to be the column name of actual choice.

\code{none} to specify the column name of the \code{none}
alternative in the validation task.

Please be aware of the following 2x2 table regarding the coding of buy and
no-buy choices:

\tabular{crcc}{
\tab \tab  Predicted           \tab    \cr
Observed \tab \tab  Buy \tab No-Buy  \cr
\tab Buy \tab A \tab B  \cr
\tab No-Buy \tab C \tab D  \cr
}
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  varskeep = "group",
  choice = "hot"
)

# precision - without group argument defined
precision(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  none = none
)

# precision - with group argument defined
precision(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  none = none,
  group = group
)

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R:
Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=accuracy]{accuracy()}}
\code{\link[=f1]{f1()}}
\code{\link[=recall]{recall()}}
\code{\link[=specificity]{specificity()}}
}
}
