% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recall.R
\name{recall}
\alias{recall}
\title{Function to calculate recall of confusion matrix for general product demand}
\usage{
recall(data, group, opts, choice, none)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{recall} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice in the validation/holdout task.}

\item{none}{Column name of none alternative.}
}
\value{
a tibble
}
\description{
\code{recall} is one of the 5 metrics of the confusion matrix
and is defined as \eqn{\frac{TP}{TP + FN}}, where TP =
True Positives, FN = False Negatives (see, e.g., Burger, 2018).
}
\details{
The current logic of \code{recall} is to determine whether a binary coded choice
is correctly predicted by the model. To use the function, the validation/holdout task
must include a \code{none} alternative.
One possible application is, for example, whether a buy or a no-buy choice
has been correctly predicted. For example, suppose you have three alternatives plus
a \code{none} alternative and want to check whether a buy or no-buy was
correctly predicted. This function can be useful if you, for example,
test whether your model significantly overestimates or underestimates
a purchase likelihood.

\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()}
function.

\code{group} optional grouping variable(s), if results should be displayed by
different group(s). Has to be column name of variables in \code{data}.

\code{opts} is required to specify the different alternatives in the
validation/holdout task (also includes the \code{none} alternative).
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice in the validation/holdout task.
Input of \code{choice} has to be column name of actual choice.

\code{none} is required to specify column name of the \code{none} alternative in the
validation/holdout task.

Please be aware about the following 2x2 table regarding coding of buy and
no-buy choice:

\tabular{crcc}{
   \tab \tab  Predicted           \tab    \cr
Observed \tab \tab  Buy \tab No-Buy  \cr
 \tab Buy \tab A \tab B  \cr
 \tab No-Buy \tab C \tab D  \cr
}
}
\examples{

library(validateHOT)

HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)
# recall - without group argument defined
recall(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  none = None
)

# recall - with group argument defined
recall(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  none = None,
  group = Group
)

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R:
Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=accuracy]{accuracy}}
\code{\link[=f1]{f1}}
\code{\link[=precision]{precision}}
\code{\link[=specificity]{specificity}}
}
}
