% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShareOfPref.R
\name{shareofpref}
\alias{shareofpref}
\title{Share of Preferences of Options included in HOT}
\usage{
shareofpref(data, id, Group = NULL, opts)
}
\arguments{
\item{data}{a data frame}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"shareofpref"} by group.}

\item{opts}{vector of column indexes of the alternatives included in the
validation/ holdout task.}
}
\value{
a data frame; a list if \code{Group} is specified
}
\description{
Function to measure the share of preferences of each option in the validation task
}
\details{
Share of Preference provides the aggrgated share of each alternative in the
validation/ holdout task as well as the lower and upper confidence interval
of each alternative which is calculated according to the following formula
\eqn{mean +/- 1.96 x \frac{sd}{\sqrt(n)}} (Orme, 2020, p. 94).

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{id} needs to be the column index of the id (unique for each participant)
in \code{data}.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a vector of the column index of \code{Group}.

\code{opts} is needed to specify the different alternatives in the validation/ holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

shareofpref(data = HOT, id = 1, opts = c(2:9))
}

\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

shareofpref(data = HOT, id = 1, opts = c(2:9), Group = 10)
}

}
\references{
{

Orme, B. K. (2020). \emph{Getting Started with Conjoint Analysis:
Strategies for Product Design and Pricing Research}. 4th edition.
Manhattan Beach, CA: Research Publishers LLC.

}
}
