% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame including Holdout Options and actual \code{"choice"} and \code{"Group"} if optional argument is defined}

\item{id}{column index}

\item{Group}{optional grouping variable to get accuracy by group}

\item{opts}{column indexes of the options included in the validation task}

\item{choice}{column index of the actual choice}

\item{None}{column index}
}
\value{
a data frame
}
\description{
test
}
\examples{
library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

specificity(data = HOT, id = 1, opts = c(2:9), choice = 10, None = 9)

library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

specificity(data = HOT, id = 1, Group = 10, opts = c(2:9), choice = 11, None = 9)

}
