% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf_ladder.R
\name{turf_ladder}
\alias{turf_ladder}
\title{Function to run Total Unduplicated Reach and Frequency ladder technique}
\usage{
turf_ladder(data, opts, none, approach = c("thres", "fc"), fixed = NULL)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{opts}{Column names of the alternatives included in the assortment.}

\item{none}{Column name of none / threshold alternative.}

\item{approach}{A character whether to run the first-choice approach
(\code{'fc'}) or threshold approach (\code{'thres'}).}

\item{fixed}{An optional vector to determine alternatives / items that have
to be included in the assortment.}
}
\value{
data frame
}
\description{
T(otal) U(nduplicated) R(each) and F(requency) is a "product line extension
model" (Miaoulis et al., 1990, p. 29). This function calculates the so-called
TURF ladder by starting with one alternative (or a fixed set of alternatives)
and then subsequently adds the next best alternative to add for increasing
the reach.
}
\details{
\code{data} a \code{data.frame} object containing the alternatives that
should be tested.

\code{opts} to specify the different alternatives in the product assortment that
should be considered. Input of \code{opts} has to be the column names of variables
in \code{data}.

\code{none} to specify the column name of the \code{none} alternative in \code{data}
that needs to be exceeded.

\code{fixed} has to be a vector of variables that are fixed in the
assortment, i.e., they must be part of the assortment.

\code{approach} character defining whether first choice \code{approach = 'fc'} or
threshold \code{approach = 'thres'} should be applied for running \code{turf()}. If
\code{approach = 'fc'}, participants are considered to be reached, if their
alternative with the highest utility is included in the assortment and this
alternative's utility is larger than the threshold's utility (Chrzan & Orme,
2019, p. 111).
On the contrary, if \code{approach = 'thres'}, participants are considered
to be reached, if the utility of one product is higher than the one of the
\code{none} alternative (Chrzan & Orme, 2019, p. 112).
If \code{approach = 'fc'}, \code{reach} equals \code{freq} since participants have at
maximum their most preferred alternative that exceeds the \code{none}
alternative.

Returns a data frame with the size of the portfolio, the reach score, the
reach add (i.e., additional reach for adding this new alternative to the
portfolio), frequency, the frequency add (i.e., additional frequency for
adding this new alternative to the portfolio), and the coding of which
alternatives are present in this portfolio (indicated by \code{1}).
}
\examples{

# turf ladder, no fixed alternatives
turf_ladder(
  data = maxdiff,
  opts = c(option_01:option_16),
  none = none,
  approach = "thres"
)

# turf ladder, alternative and 4 and 5 fixed
turf_ladder(
  data = maxdiff,
  opts = c(option_01:option_16),
  none = none,
  fixed = c("option_04", "option_05"),
  approach = "thres"
)

}
\references{
{

Chrzan, K., & Orme, B. K. (2019). \emph{Applied MaxDiff: A Practitioner’s
Guide to Best-Worst Scaling} Provo, UT: Sawtooth Software.

Miaoulis, G., Parsons, H., & Free, V. (1990). Turf: A New Planning Approach
for Product Line Extensions. \emph{Marketing Research 2} (1): 28-40.

}
}
