% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validateHOT-package.R
\docType{package}
\name{validateHOT}
\alias{validateHOT-package}
\alias{validateHOT}
\title{validateHOT: validateHOT - An R Package to validate holdout tasks and run market simulations}
\description{
validateHOT is a package to validate validation tasks in preference
measurement techniques, run market simulations, and convert raw utilities
into scores that are easier to interpret.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/JoshSchramm94/validateHOT}
  \item Report bugs at \url{https://github.com/JoshSchramm94/validateHOT/issues}
}

}
\author{
\strong{Maintainer}: Joshua Schramm \email{joshua.schramm@mail.de} (\href{https://orcid.org/0000-0001-5602-4632}{ORCID})

Other contributors:
\itemize{
  \item Marcel Lichers (\href{https://orcid.org/0000-0002-3710-2292}{ORCID}) [contributor]
}

}
