% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zc_diffs.R
\name{zc_diffs}
\alias{zc_diffs}
\title{Zero-Centered Diffs for (A)CBC}
\usage{
zc_diffs(
  data,
  group = NULL,
  attrib,
  coding,
  interpolate.levels = NULL,
  res = c("agg", "ind"),
  none = NULL
)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s).}

\item{attrib}{A list that specifies the attribute levels for each attribute.}

\item{coding}{A vector of the coding of each attribute, '0' = part-worth
coding, '1' = linear coding.}

\item{interpolate.levels}{A list of the attribute levels that should
be interpolated. These have to be the same as provided to Sawtooth Software.
Please make sure to provide the whole list. Only has to be specified for the
variables that are coded as '1' (linear).}

\item{res}{A vector indicating whether individual zero-centered diffs (\code{ind}) or
aggregated (\code{agg}) zero-centered diffs should be returned.}

\item{none}{A vector whether \code{none} option was included.}
}
\value{
a tibble
}
\description{
Zero-Centered Diffs for (A)CBC
}
\details{
\code{zc_diffs} converts raw utilities of a CBC or an ACBC to
zero-centered diffs (Orme, 2020, p. 78). This allows for comparison between
the attributes.

\code{data} has to be a data frame with the attributes. Attribute levels need
to be the raw utilities.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{attrib} specifies the attribute levels for each alternative.
Input for \code{attrib} has to be a list. Needs to specify the column names or
column indexes of the attribute levels.

\code{coding} has to be specified  to indicate the attribute coding. \code{0}
to indicated part-worth coding, \code{1} for linear coding.

\code{interpolate.levels} has to be specified for linear coded variables.
If scaled or centered values were used for hierarchical bayes (HB)
estimation, these have to be specified in this case.
All values have to be specified. For example, if one linear coded attribute
had 5 levels, all 5 levels have to be inserted.

\code{res} specifies whether results should be aggregated across all participants
or across \code{group} (\code{res} needs to be set to {agg}) or if it scores
should be converted for individuals only

\code{none} specifies whether none option was included or not, if yes,
column name or column index of \code{none} needs to be specified. If no
\code{none} option was included, leave it empty.
}
\examples{
\dontrun{
zc_diffs(
  data = CBC,
  attrib = list(
    c("Att1_Lev1", "Att1_Lev2", "Att1_Lev3", "Att1_Lev4", "Att1_Lev5"),
    c("Att2_Lev1", "Att2_Lev2", "Att2_Lev3", "Att2_Lev4", "Att2_Lev5"),
    c("Att3_Lev1", "Att3_Lev2", "Att3_Lev3", "Att3_Lev4", "Att3_Lev5", "Att3_Lev6", "Att3_Lev7")
  ),
  coding = c(0, 0, 0),
  none = "none"
)

zc_diffs(
  data = CBC_lin,
  attrib = list(
    c("Att1_Lev1", "Att1_Lev2", "Att1_Lev3", "Att1_Lev4", "Att1_Lev5"),
    c("Att2_Lev1", "Att2_Lev2", "Att2_Lev3", "Att2_Lev4", "Att2_Lev5"),
    c("Att3_Lin")
  ),
  coding = c(0, 0, 1),
  interpolate.levels = list(c(10, 20, 30, 40, 50, 60, 70)),
  none = "none"
)
}

}
\references{
{

Orme, B. K. (2020). \emph{Getting Started with Conjoint Analysis: Strategies for Product Design and Pricing Research}.
4th edition. Manhattan Beach, CA: Research Publishers LLC.

}
}
\seealso{
{
\code{\link[=att_imp]{att_imp}} for attribute importance scores for (A)CBC
\code{\link[=prob_scores]{prob_scores}} for probability scores for MaxDiff
}
}
