% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zero_anchored.R
\name{zero_anchored}
\alias{zero_anchored}
\title{Function to calculate zero-anchored interval scores for (anchored) MaxDiff}
\usage{
zero_anchored(data, group = NULL, items, res = c("agg", "ind"), anchor = NULL)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s).}

\item{items}{Vector that specifies the items.}

\item{res}{A vector indicating whether individual shares (\code{ind}) or
aggregated (\code{agg}) shares should be returned.}

\item{anchor}{An optional variable to specify anchor variable.}
}
\value{
a tibble
}
\description{
Function to calculate zero-anchored interval scores for (anchored) MaxDiff
}
\details{
\code{zero_anchored} converts raw utilities of a MaxDiff to zero-anchored interval scores that have a range of 100.

For anchored MaxDiff the anchor is set to 0. More information can be obtained here: https://sawtoothsoftware.com/help/lighthouse-studio/manual/analysis-manager-maxdiff-export-settings.html

\code{data} has to be a data frame with the attributes. Items have
to be the raw utilities.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{items} specifies the items of the MaxDiff.
Input for \code{items} has to be variable names.

\code{res} specifies whether results should be aggregated across all participants
or across \code{group} (\code{res} needs to be set to \code{agg}) or if scores
should be converted for individuals only.

\code{anchor} only needs to be specified if anchored MaxDiff is applied.
Input for \code{anchor} has to be variable names.
}
\examples{
\dontrun{

# zero-anchored interval scores for unanchored MaxDiff - without Group defined
zero_anchored(
  data = MaxDiff,
  items = c(Option_01:Option_16),
  res = "agg"
)

# zero-anchored interval scores for unanchored MaxDiff - with Group defined
zero_anchored(
  data = MaxDiff,
  group = Group,
  items = c(Option_01:Option_16),
  res = "agg"
)


# zero-anchored interval scores for anchored MaxDiff - without Group defined
zero_anchored(
  data = MaxDiff,
  items = c(Option_01:none),
  anchor = none,
  res = "agg"
)

# zero-anchored interval scores for anchored MaxDiff - with Group defined
zero_anchored(
  data = MaxDiff,
  group = Group,
  items = c(Option_01:none),
  anchor = none,
  res = "agg"
)
}

}
\references{
{

Chrzan, K., & Orme, B. K. (2019). \emph{Applied MaxDiff: A Practitioner’s
Guide to Best-Worst Scaling} Provo, UT: Sawtooth Software.

}
}
\seealso{
{
\code{\link[=att_imp]{att_imp}} for attribute importance scores for (A)CBC
\code{\link[=prob_scores]{prob_scores}} for probability scores for MaxDiff
\code{\link[=zc_diffs]{zc_diffs}} for zero-center diff scores for (A)CBC
}
}
