% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame including alternatives in the validation task and actual \code{"choice"}. \code{"None"} alternative needs to be included.}

\item{id}{column index of \code{"id"}}

\item{Group}{optional grouping variable to get accuracy by group}

\item{opts}{column indexes of the alternatives included in the validation task}

\item{choice}{column index of the actual choice}

\item{None}{column index of None alternative}
}
\value{
a data frame
}
\description{
accuracy is defined as number of correct predicted participants divided by the total number of predictions
}
\details{
The current logic of \code{"accuracy"} is to provide whether a binary coded event is correctly predicted. To use the function a \code{"None"} alternative in the script.
One potential usage is, for example, whether a buy or a no-buy condition was predicted correctly. For example, you have three alternatives plus a \code{"None"} alternative and you want to check whether a buy or no-buy was correclty predicted.

\code{data} needs to be a data frame including the alternatives shown in the validation/holdout
task. Can be created using the createHOT function.

\code{id} needs to be the column index of the id (unique for each participant) in data frame.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a column index.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).

\code{choice} specifies the column index of the actual choice.
Input of opts \code{choice} needs to be the column index of actual choice.

\code{None} specifies the column index of the \code{None} alternative in the
validation/holdout task. Needs to be specified for \code{accuracy()}.
}
\examples{
library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

accuracy(data = HOT, id = 1, opts = c(2:9), choice = 10, None = 9)

library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

accuracy(data = HOT, id = 1, Group = 10, opts = c(2:9), choice = 11, None = 9)

}
\seealso{
{
\code{\link[=createHOT]{createHOT}}
}
}
\concept{family name}
