% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accuracy.R
\name{accuracy}
\alias{accuracy}
\title{accuracy}
\usage{
accuracy(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame including alternatives in the validation task and actual \code{"choice"}. \code{"None"} alternative needs to be included.}

\item{id}{column index of \code{"id"}}

\item{Group}{optional grouping variable to get accuracy by group}

\item{opts}{column indexes of the alternatives included in the validation task}

\item{choice}{column index of the actual choice}

\item{None}{column index of None alternative}
}
\value{
a data frame
}
\description{
accuracy is defined as number of correct predicted participants divided by the total number of predictions
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff")
accuracy(data = HOT, id = 1, opts = c(2:9), choice = 10, None = 9)


library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff", varskeep = 21)
accuracy(data = HOT, id = 1, Group = 10, opts = c(2:9), choice = 11, None = 9)


}
