% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Accuracy.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(data, group, opts, choice, none)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{"accuracy"} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice.}

\item{none}{Column name of none alternative.}
}
\value{
a tibble
}
\description{
\code{accuracy} is one of the 5 metrics of the confusion matrix
and is defined as number of correctly predicted participants divided by
the total number of predictions. See, for example, Burger (2018):
\eqn{\frac{TP + TN}{TP + FP + TN + FN}}, where TP =
True Positives, TN = True Negatives, FP = False Positives, and FN =
False Negatives.
}
\details{
The current logic of \code{"accuracy"} is to determine whether a binary coded
is correctly predicted by the model. To use the function a \code{"none"}
alternative has to be included in the validation/holdout task.
One potential usage is, for example, whether a buy or a no-buy condition
was predicted correctly. For example, assume you have three alternatives plus
a \code{"none"} alternative and you want to check whether a buy or no-buy was
correctly predicted. This function can be helpful when you test, for example,
if your model significantly overestimates or underestimates, for
example, a purchase likelihood.


\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()}
function.

\code{group} optional grouping variable(s), if results should be displayed by
different groups. Has to be column name(s) of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the
validation/holdout task (also includes the \code{none} alternative).
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of \code{choice} has to be column name of actual choice.

\code{none} to specify column name of the \code{none} alternative in the
validation/holdout task.

Please be aware about the following 2x2 table regarding coding of buy and
no-buy choice

\tabular{crcc}{
   \tab \tab  Predicted           \tab    \cr
Observed \tab \tab  Buy \tab No-buy  \cr
 \tab Buy \tab A \tab B  \cr
 \tab No-Buy \tab C \tab D  \cr
}
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

# accuracy ungrouped
accuracy(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  none = None
)

# accuracy by group
accuracy(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  none = None,
  group = Group
)
}

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R:
Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=f1]{f1}}
\code{\link[=precision]{precision}}
\code{\link[=recall]{recall}}
\code{\link[=specificity]{specificity}}
}
}
