% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_imp.R
\name{att_imp}
\alias{att_imp}
\title{Function to calculate attributes' importance score of (A)CBCs}
\usage{
att_imp(
  data,
  group = NULL,
  attrib,
  coding,
  interpolate.levels = NULL,
  res = c("agg", "ind")
)
}
\arguments{
\item{data}{A data.frame object.}

\item{group}{Optional variable to to get \code{att_imp()} by group(s).}

\item{attrib}{A list that specifies the attribute levels for each attribute.}

\item{coding}{A vector of the coding of each attribute, \code{0} = part-worth
coding, \code{1} = linear coding, or \code{2} = piecewise coding.}

\item{interpolate.levels}{A list to specify the values of variables that
are linear coded. These have to be the same as specified in model estimation
(e.g., if you center attribute levels before estimation, insert the centered
levels). Please make sure to provide the whole list (i.e., all levels). Only
has to be specified for the variables that are coded as \code{1} (linear).}

\item{res}{A vector indicating whether individual results (\code{ind}) or
aggregated (\code{agg}) results should be returned.}
}
\value{
a tibble
}
\description{
Function to calculate attributes' importance score of (A)CBCs
}
\details{
\code{att_imp()} converts raw utilities of a CBC or an ACBC to relative
importance scores (see, Orme, 2020, p. 80, for more information).

\code{data} A data frame with all attributes and the corresponding
levels. Attribute levels need to be the raw utilities of hierarchical Bayes
estimation.

\code{group} optional grouping variable, if results should be displayed by
different groups.

\code{attrib} specifies the attribute levels for each alternative.
Input for \code{attrib} has to be a list. Needs to specify the column names
or column indexes of the attribute levels.

\code{coding} is required to indicate the attributes' coding. \code{0}
to indicate part-worth coding, \code{1} for linear coding, or \code{2} for
piecewise coding.

\code{interpolate.levels} is required for linear coded variables.
If scaled or centered values were used for hierarchical Bayes
estimation, these need to be defined in \code{att_imp()}.
All values have to be specified. For example, if one linear coded attribute
has 5 levels, all 5 levels have to be inserted.

\code{res} specifies whether results should be aggregated across all
participants or across \code{group} (\code{res} needs to be set to
\code{agg}) or if scores should be converted for individuals only (\code{ind}).
}
\examples{

att_imp(
  data = cbc,
  attrib = list(
    paste0("att1_lev", c(1:3)),
    paste0("att2_lev", c(1:2)),
    paste0("att3_lev", c(1:4)),
    paste0("att4_lev", c(1:4)),
    paste0("att5_lev", c(1:2)),
    paste0("att6_lev", c(1:4)),
    paste0("att7_lev", c(1:6)),
    paste0("att8_lev", c(1:6)),
    paste0("price_", c(1:6))
  ),
  coding = c(rep(0, times = 9)),
  res = "agg"
)

att_imp(
  data = cbc_linear,
  attrib = list(
    paste0("att1_lev", c(1:3)),
    paste0("att2_lev", c(1:2)),
    paste0("att3_lev", c(1:4)),
    paste0("att4_lev", c(1:4)),
    paste0("att5_lev", c(1:2)),
    paste0("att6_lev", c(1:4)),
    paste0("att7_lev", c(1:6)),
    paste0("att8_lev", c(1:6)),
    "price"
  ),
  coding = c(rep(0, times = 8), 1),
  interpolate.levels = list(c(seq(from = 175.99, to = 350.99, by = 35))),
  res = "agg"
)

}
\references{
{

Orme, B. K. (2020). \emph{Getting Started with Conjoint Analysis: Strategies
for Product Design and Pricing Research}. 4th edition. Manhattan Beach,
CA: Research Publishers LLC.

}
}
\seealso{
{
\code{\link[=prob_scores]{prob_scores()}} for probability scores for MaxDiff
\code{\link[=zero_anchored]{zero_anchored()}} for zero-anchored interval
scores for MaxDiff
\code{\link[=zc_diffs]{zc_diffs()}} for zero-centered diff scores for (A)CBC
}
}
