% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/att_imp.R
\name{att_imp}
\alias{att_imp}
\title{Title}
\usage{
att_imp(data, group = NULL, attrib, coding, interpolate.levels = NULL)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)}

\item{attrib}{specifies the attribute levels for each attribute}

\item{coding}{a vector of the coding of each attribute, '0' = part-worth
coding, '1' = linear coding}

\item{interpolate.levels}{a list of the attribute levels that should
be interpolated. These have to be the same as provided to Sawtooth Software.
Please make sure to provide the whole list. Only has to be specified for the
variables that are coded as '1' (linear)}
}
\value{
a tibble
}
\description{
Title
}
\examples{
\dontrun{
att_imp(data = CBC,
attrib = list(c("Att1_Lev1", "Att1_Lev2", "Att1_Lev3", "Att1_Lev4", "Att1_Lev5"),
c("Att2_Lev1", "Att2_Lev2", "Att2_Lev3", "Att2_Lev4", "Att2_Lev5"),
c("Att3_Lev1", "Att3_Lev2", "Att3_Lev3", "Att3_Lev4", "Att3_Lev5", "Att3_Lev6", "Att3_Lev7")),
coding = c(0, 0, 0))

att_imp(data = CBC_lin,
attrib = list(c("Att1_Lev1", "Att1_Lev2", "Att1_Lev3", "Att1_Lev4", "Att1_Lev5"),
c("Att2_Lev1", "Att2_Lev2", "Att2_Lev3", "Att2_Lev4", "Att2_Lev5"),
c("Att3_Lin")),
coding = c(0, 0, 1),
interpolate.levels = list(c(10, 20, 30, 40, 50, 60, 70)))

}

}
