% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createHOT.R
\name{createHOT}
\alias{createHOT}
\title{Preparing Holdout task and creating utilities}
\usage{
createHOT(
  data,
  id,
  None = NULL,
  prod,
  prod.levels,
  interpolate.levels = NULL,
  piece.p = NULL,
  lin.p = NULL,
  coding = NULL,
  method = c("ACBC" | "CBC" | "MaxDiff"),
  varskeep = NULL,
  choice
)
}
\arguments{
\item{data}{a data frame}

\item{id}{the column index of \code{id} in \code{data}}

\item{None}{the column index of \code{None} in \code{data}; if of \code{None} is not included, leave empty}

\item{prod}{number of options in the Holdout task without the \code{None} option, must be numeric}

\item{prod.levels}{define the attribute levels of the products, must be a list}

\item{interpolate.levels}{xyz}

\item{piece.p}{xyz}

\item{lin.p}{xyz}

\item{coding}{xyz}

\item{method}{specify the \code{method} your study; needs to be one of the following: MaxDiff, CBC, or ACBC}

\item{varskeep}{variables that should be kept in the data frame, use column index}

\item{choice}{actual choice in the Holdout task}
}
\value{
a data frame
}
\description{
Function used to create utilities for validation task.
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(
  data = MaxDiff, None = 19,
  id = 1, prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  choice = 20, method = "MaxDiff"
)

}
