% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f1.R
\name{f1}
\alias{f1}
\title{Function to calculate F1-Score of confusion matrix for general product demand}
\usage{
f1(data, group, opts, choice, none)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{f1()} by group(s).}

\item{opts}{Column names of the alternatives in the validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}

\item{none}{Column name of none alternative.}
}
\value{
a tibble
}
\description{
F1-Score is one of the 5 metrics of the confusion matrix
and is defined as \eqn{\frac{2 * precision * recall}{precision + recall}} or
stated differently by Burger (2018) \eqn{\frac{2TP}{2TP + FP + FN}},
where TP = True Positives, FP = False Positives, and FN = False Negatives.
}
\details{
The current logic of \code{f1()} is to determine whether a binary coded
choice is correctly predicted by the model. To use the function, the
validation task must include a \code{none} alternative.
One possible application is, for example, whether a buy or a no-buy choice
has been correctly predicted. For example, suppose you have three
alternatives plus a \code{none} alternative and want to check whether a buy
or no-buy was correctly predicted. This function can be useful if you, for
example, test whether your model significantly overestimates or
underestimates a purchase likelihood.

\code{data} a data frame including the alternatives shown in the validation task.
Can be created using the \code{create_hot()} function.

\code{group} optional grouping variable(s), if results should be displayed by
different groups. Has to be column name(s) of variables in \code{data}.

\code{opts} to specify the different alternatives in the
validation task (also includes the \code{none} alternative).
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice in the validation
task. Input of \code{choice} has to be column name of actual choice.

\code{none} to specify column name of the \code{none}
alternative in the validation task.

Please be aware about the following 2x2 table regarding coding of buy and
no-buy choice:

\tabular{crcc}{
\tab \tab  Predicted           \tab    \cr
Observed \tab \tab  Buy \tab No-Buy  \cr
\tab Buy \tab A \tab B  \cr
\tab No-Buy \tab C \tab D  \cr
}
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  varskeep = "group",
  choice = "hot"
)

# f1 - without group argument defined
f1(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  none = none
)

# f1 - with group argument defined
f1(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  none = none,
  group = group
)

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R:
Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=accuracy]{accuracy()}}
\code{\link[=precision]{precision()}}
\code{\link[=recall]{recall()}}
\code{\link[=specificity]{specificity()}}
}
}
