% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{f1}
\alias{f1}
\title{F1-Score}
\usage{
f1(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame with all relevant variables.}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"f1"} by group}

\item{opts}{vector of column indexes of the alternatives included in the
validation/holdout task.}

\item{choice}{vector of column index of the actual choice.}

\item{None}{vector of column index of None alternative.}
}
\value{
a data frame
}
\description{
F1 is one of the 5 metrics of the confusion matrix
and is defined as \eqn{\frac{2 * precision * recall}{precision + recall}} or stated
differently by Burger (2018) \eqn{\frac{2TP}{2TP + FP + FN}}, where TP =
True Positives, FP = False Positives, and FN = False Negatives.
}
\details{
The current logic of \code{"f1"} is to provide whether a binary coded event is correctly predicted.
To use the function a \code{"None"} alternative needs to be in the script.
One potential usage is, for example, whether a buy or a no-buy condition
was predicted correctly. For example, you have three alternatives plus
a \code{"None"} alternative and you want to check whether a buy or no-buy was
correctly predicted. This function can be helpful when you test whether or
not your model significantly overestimates or underestimates, for example, a purchase likelihood.

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{id} needs to be the column index of the id (unique for each participant)
in \code{data}.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a vector of the column index of \code{Group}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).

\code{choice} specifies the column index of the actual choice.
Input of opts \code{choice} needs to be the column index of actual choice.

\code{None} specifies the column index of the \code{None} alternative in the
validation/holdout task. Needs to be specified for \code{f1()}.
}
\examples{
library(validateHOT)
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

f1(data = HOT, id = 1, opts = c(2:9), choice = 10, None = 9)

library(validateHOT)
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

f1(data = HOT, id = 1, Group = 10, opts = c(2:9), choice = 11, None = 9)

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R: Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=accuracy]{accuracy}}
\code{\link[=precision]{precision}}
\code{\link[=recall]{recall}}
\code{\link[=specificity]{specificity}}
}
}
