% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/F1.R
\name{f1}
\alias{f1}
\title{F1-Score}
\usage{
f1(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame including alternatives in the validation task and actual \code{"choice"}. \code{"None"} alternative needs to be included.}

\item{id}{column index of \code{"id"}}

\item{Group}{optional grouping variable to get F1-score by group}

\item{opts}{column indexes of the alternatives included in the validation task}

\item{choice}{column index of the actual choice}

\item{None}{column index of None alternative}
}
\value{
a data frame
}
\description{
Calculates the F1-Score.  F1-Score is calculated by the following formula \eqn{2 * precision * recall / (precision + recall)}
}
\details{
To get the \code{"f1"} score of the validation task, you have to specify the \code{"None"} alternative in the script. The function will calculate \code{"f1"} based on whether or not, for example, a buy or a no-buy was correctly predicted.
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(
  data = MaxDiff, None = 19,
  id = 1, prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  choice = 20, method = "MaxDiff"
)
f1(data = HOT, id = 1, opts = c(2:9), choice = 10, None = 9)

library(ValiDatHOT)
data(MaxDiff)
createHOT(
  data = MaxDiff, None = 19,
  id = 1, prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  choice = 20, method = "MaxDiff", varskeep = 21
)
f1(data = HOT, id = 1, Group = 10, opts = c(2:9), choice = 11, None = 9)

}
