% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqAssort.R
\name{freqassort}
\alias{freqassort}
\title{Function to calculate averaged number of times a participant is reached by assortment}
\usage{
freqassort(data, group, none, opts)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{freqassort} by group(s).}

\item{none}{Column name of none / threshold alternative.}

\item{opts}{Column names of the alternatives included in the assortment.}
}
\value{
a tibble
}
\description{
Frequency function of T(otal) U(nduplicated) R(each) and F(requency)
analysis to measure the average time a consumer
is reached with a particular product bundle assortment.
}
\details{
Frequency calculates the average times a consumer would be reached with the
tested product assortment. The current logic of \code{freqassort}
is that the utility of an alternative has to exceed a threshold. In the case
of \code{freqassort} this threshold is referred to the \code{none} argument
in \code{data}.
The frequency is calculated based on the 'threshold' approach, i.e.,
each alternative that exceeds utility of \code{none} alternative is
considered as, for example, purchase option.

\code{data} has to be a data frame including the alternatives that should
be tested.

\code{group} optional grouping variable, if results should be displayed
by different groups. Has to be column name of variables in \code{data}.

\code{opts} defines product assortment that should be considered.
Input of \code{opts} has to be column names of variables in \code{data}.

\code{none} to specify column name of the \code{none} alternative (i.e.,
threshold variable).
}
\examples{

library(validateHOT)

HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# freqassort - without group argument defined
freqassort(
  data = HOT,
  opts = c(Option_1, Option_2, Option_6),
  none = None
)

# freqassort - with group argument defined
freqassort(
  data = HOT,
  opts = c(Option_1, Option_2, Option_6),
  none = None,
  group = Group
)

}
