% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqAssort.R
\name{freqassort}
\alias{freqassort}
\title{Averaged number of times a person is reached by a specific assortment of bundles}
\usage{
freqassort(
  data,
  id,
  Group = NULL,
  None,
  method = c("threshold" | "First Choice"),
  bundles
)
}
\arguments{
\item{data}{data data frame with all relevant variables.}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"accuracy"} by group.}

\item{None}{vector of column index of None alternative.}

\item{method}{character variable that needs to be specified to either \code{"First choice"} or \code{"threshold"}, please see Details}

\item{bundles}{vector of column indexes of the bundles that should be included in the assortment.}
}
\value{
a data frame
}
\description{
Frequency function of TURF analysis to measure the average time a consumer
is reached with a specific product bundle assortment. Can either be calculated
using 'First Choice' or 'threshold' rule. See Details section for more
information and specifying the data in the correct way.
}
\details{
Please specify the method you are using. The method can be either set
to \code{"First choice"} (only the bundle with highest utility is marked as purchase option if
its above \code{"None"} utility) or \code{"threshold"}
(all bundles with utility above \code{"None"} utility are marked as purchase option).
}
\examples{
library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

freqassort(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "threshold")
freqassort(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "First Choice")

library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

freqassort(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "threshold", Group = 10)
freqassort(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "First Choice", Group = 10)

}
