% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqassort.R
\name{freqassort}
\alias{freqassort}
\title{Averaged number of times a person is reached by a specific assortment of bundles}
\usage{
freqassort(
  data,
  id,
  Group = NULL,
  None,
  method = c("threshold" | "First Choice"),
  bundles
)
}
\arguments{
\item{data}{data frame including alternatives in the validation task and actual \code{"choice"}. \code{"None"} alternative needs to be included.}

\item{id}{column index of \code{"id"}}

\item{Group}{optional grouping variable to get results by group}

\item{None}{column index of None alternative}

\item{method}{either \code{"First choice"} or \code{"threshold"}, please see Details}

\item{bundles}{column indexes of the bundles that should be included in the assortment}
}
\value{
a data frame
}
\description{
Frequency function of TURF analysis to measure the average time a consumer
is reached with a specific product bundle assortment. Can either be calculated
using 'First Choice' or 'threshold' rule. See Details section for more
information and specifying the data in the correct way.
}
\details{
Please specify the method you are using. The method can be either set
to \code{"First choice"} (only the bundle with highest utility is marked as purchase option if
its above \code{"None"} utility) or \code{"threshold"}
(all bundles with utility above \code{"None"} utility are marked as purchase option).
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff")
freqassort(data = HOT, id = 1, bundles = c(2,3,7), None = 9, method = "threshold")
freqassort(data = HOT, id = 1, bundles = c(2,3,7), None = 9, method = "First Choice")

library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff", varskeep = 21)
freqassort(data = HOT, id = 1, bundles = c(2,3,7), None = 9, method = "threshold", Group = 10)
freqassort(data = HOT, id = 1, bundles = c(2,3,7), None = 9, method = "First Choice", Group = 10)

}
