% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqAssort.R
\name{freqassort}
\alias{freqassort}
\title{Averaged number of times a person is reached by a specific assortment of bundles}
\usage{
freqassort(data, group, none, opts)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)
to get \code{"freqassort"} by group(s)}

\item{none}{column name of none / threshold alternative}

\item{opts}{column names of the alternatives included in the assortment}
}
\value{
a tibble
}
\description{
Frequency function of T(otal) U(nduplicated) R(each) and F(requency)
analysis to measure the average time a consumer
is reached with a specific product bundle assortment. \code{"freqassort"} calculates
the frequency based on the 'threshold' approach, meaning each alternative that exceeds
utility of \code{none} alternative is considered as, for example, purchase option.
}
\details{
Frequency calculates the average times a consumer would be reached with the
product assortment you are testing. The current logic of \code{freqassort()}
is that the utility of an alternative has to exceed a threshold. In the case of \code{freqassort()}
this threshold is referred to the \code{none} argument in \code{data}.

\code{data} has to be a data frame including the alternatives that should be tested

\code{group} optional grouping variable, if results should be displayed by different conditions.
Has to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the
product assortment that should be considered.
Input of \code{opts} has to be column names of variables in \code{data}.

\code{none} to specify column name of the \code{none} alternative in the
validation/holdout task.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# freqassort ungrouped
freqassort(data = HOT,
           opts = c(Option_1, Option_2, Option_6),
           none = None)

# freqassort grouped
freqassort(data = HOT,
           opts = c(Option_1, Option_2, Option_6),
           none = None,
           group = Group)
}

}
