% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitrate.R
\name{hitrate}
\alias{hitrate}
\title{Function to calculate hit rate of validation task}
\usage{
hitrate(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{hitrate()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}
}
\description{
\code{hitrate()} measures number of times a choice was correctly
predicted in a validation task.
}
\details{
\code{hitrate()} measures number of times a participant's choice was correctly
predicted by the model.
Output contains the following 5 metrics:
\itemize{
\item \code{hr} hit rate (number of correctly predicted
choices) in percentages
\item \code{se} standard error in percentages
\item \code{chance} chance level of correctly predicted choices by just
guessing
\item \code{cor} absolute number of correctly predicted choices
\item \code{n} total number of choices
}

\code{data} a data.frame object including the alternatives shown in the
validation task. Can be created using the \code{create_hot()} function.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{opts} is required to specify the different alternatives in the validation
task. Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice in the validation
task. Input of opts \code{choice} has to be column name of actual choice.
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  varskeep = "group",
  choice = "hot"
)

# hit rate - without group argument defined
hitrate(
  data = hot,
  opts = c(option_1:none),
  choice = choice
)

# hit rate - with group argument defined
hitrate(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  group = group
)

}
