% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitrate.R
\name{hitrate}
\alias{hitrate}
\title{Hit Rate}
\usage{
hitrate(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{frame with all relevant variables.}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"hitrate"} by group.}

\item{opts}{vector of column indexes of the alternatives included in the
validation/holdout task.}

\item{choice}{vector of column index of the actual choice.}
}
\value{
a data frame
}
\description{
\code{hitrate} measures number of times a choice was correctly predicted in a validation task.
}
\details{
\code{hitrate} measures number of times a participant's choice was correctly
predicted by our model.
Output contains the following 3 metrics:
\itemize{
\item \code{"chance"} chance level (\eqn{\frac{1}{number of alternatives}}) in percentage
\item \code{"no."} absolute value of correctly predicted choices
\item \code{"\%"} correctly predicted choices in percentage
}

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{id} needs to be the column index of the id (unique for each participant)
in \code{data}.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a vector of the column index of \code{Group}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).

\code{choice} specifies the column index of the actual choice.
Input of opts \code{choice} needs to be the column index of actual choice.
}
\examples{
library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

hitrate(data = HOT, id = 1, opts = c(2:9), choice = 10)

library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

hitrate(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
