% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitrate.R
\name{hitrate}
\alias{hitrate}
\title{Hit Rate}
\usage{
hitrate(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{a data frame}

\item{id}{column index of the \code{id} variable}

\item{Group}{optional grouping variable to get hit rate by group}

\item{opts}{column indexes of the options included in the validation task}

\item{choice}{column index of the actual choice}
}
\value{
a data frame
}
\description{
\code{hitrate} measures number of times a choice was correctly predicted in a validation task.
}
\details{
Output contains the chance level (\eqn{1 / # of alternatives in validation task}), the number of correct predictions and the percentage of number of correct predictions.
}
\examples{
library(ValiDatHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

hitrate(data = HOT, id = 1, opts = c(2:9), choice = 10)

library(ValiDatHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

hitrate(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
