% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hitrate.R
\name{hitrate}
\alias{hitrate}
\title{Hit Rate}
\usage{
hitrate(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{"hitrate"} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice.}
}
\value{
a tibble
}
\description{
\code{hitrate} measures number of times a choice was correctly
predicted in a validation/holdout task.
}
\details{
\code{hitrate} measures number of times a participant's choice was correctly
predicted by the model.
Output contains the following 4 metrics:
\itemize{
\item \code{"HR"} hit rate (number of correctly predicted
choices) in percentages
\item \code{"se"} standard error in percentages
\item \code{"chance"} chance level of correctly predicted choices by simply
guessing
\item \code{"cor"} absolute number of correctly predicted choices
\item \code{"n"} total number of choices
}

\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()}
function.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the
validation/holdout task (also includes the \code{none} alternative).
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of opts \code{choice} has to be column name of actual choice.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)
# hit rate ungrouped
hitrate(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice
)

# hit rate grouped
hitrate(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  group = Group
)
}

}
