% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl}
\alias{kl}
\title{Function to calculate Kullback-Leibler divergence of validation
task}
\usage{
kl(data, group = NULL, opts, choice, epsilon = NULL, log_base = NULL)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{kl()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}

\item{epsilon}{A vector of noise that should be added to \code{0} values, per
default set to \code{1e-05}.}

\item{log_base}{A character string to define the logarithm base, currently
\code{log} (default) and \code{log2} are provided.}
}
\value{
a tibble
}
\description{
Function to calculate Kullback-Leibler divergence of validation
task
}
\details{
Kullback-Leibler-Divergence calculates the divergence between the actual
choice distribution and the predicted choice distribution (Ding et al., 2011;
Drost, 2018). Currently \code{kl()} provides the deviation measured based on
\eqn{log} and \eqn{log{_2}} algorithm. \eqn{log} is set as default.

Due to Kullback-Leibler divergence's asymmetry, the output provides both
\code{KL_O_P} which is equivalent to (Observed || Predicted) and
\code{KL_P_O} which is equivalent to (Predicted || Observed).

\code{data} a \code{data.frame} object including the alternatives shown in the
validation task. It can be created using the \code{create_hot()} function.

\code{group} optional grouping variable(s) to calculate results by group(s).
different groups. Has to be the column name(s) of variables in \code{data}.

\code{opts} to specify the different alternatives in the
validation task (also includes the \code{none} alternative).

\code{choice} to specify the column of actual choice in the validation
task. The input of \code{choice} has to be the column name of actual choice.

\code{epsilon} numeric input in case of \code{0} in the numerator or
denominator. \code{0} then will be replaced by \code{epsilon}. The default value
is \code{epsilon = 1e-5} but can be adopted (see also Drost, 2018).

\code{log_base} has to be a character string, deciding which logarithm base
you want to apply to calculate Kullback-Leibler. You can choose
between \eqn{log} and \eqn{log{_2}}. The default is set to \eqn{log}.
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  choice = "hot",
  varskeep = "group"
)

# kl with log base - without group argument defined
kl(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  log_base = "log"
)

\dontrun{
# kl with log2 base - without group argument defined
kl(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  log_base = "log2"
)

# kl grouped - log + specifying epsilon
kl(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  log_base = "log",
  group = group,
  epsilon = 1e-8
)

# kl grouped - log2
kl(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  log_base = "log2",
  group = group
)
}

}
\references{
{

Ding, Min, John R. Hauser, Songting Dong, Daria Dzyabura, Zhilin Yang,
SU Chenting, and Steven P. Gaskin. (2011).
Unstructured Direct Elicitation of Decision Rules. \emph{Journal of
Marketing Research 48}(1): 116-27.
\verb{https://doi.org/10.1509/jmkr.48.1.116}.

Drost, Hajk-Georg. (2018). Philentropy: Information Theory and Distance
Quantification with R. \emph{Journal of Open Source Software 3}(26),
765, \verb{https://joss.theoj.org/papers/10.21105/joss.00765}.

}
}
