% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kl.R
\name{kl}
\alias{kl}
\title{Kullback-Leibler Divergence}
\usage{
kl(data, id, Group = NULL, opts, choice, epsilon = NULL)
}
\arguments{
\item{data}{a data frame}

\item{id}{column index of the \code{id} variable}

\item{Group}{optional grouping variable to get hit rate by group}

\item{opts}{column indexes of the options included in the holdout task}

\item{choice}{column index of the actual choice}

\item{epsilon}{noise that should be added to 0 values, per default set to 1e-05}
}
\value{
a data frame
}
\description{
Function to measure the Kullback-Leibler Divergence of a validation task
}
\details{
Additional details...
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff")
kl(data = HOT, id = 1, opts = c(2:9), choice = 10)


library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff", varskeep = 21)
kl(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
