% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{kl}
\alias{kl}
\title{Kullback-Leibler Divergence}
\usage{
kl(data, group, opts, choice, epsilon = NULL, base = NULL)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{"kl"} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice.}

\item{epsilon}{A vector of noise that should be added to 0 values, per
default set to 1e-05.}

\item{base}{A character string to define the logarithm base, currently choice
between \code{log} (default) and \code{log2}.}
}
\value{
a tibble
}
\description{
Function to measure the Kullback-Leibler Divergence of a
validation/holdout task.
}
\details{
Kullback-Leibler-Divergence which measures the divergence between the actual
choice distribution and the predicted
choice distribution (Ding et al., 2011; Drost, 2018). Currently only provides
the deviation measured based on \eqn{log} and \eqn{log{_2}}
algorithm. \eqn{log} set as default.

Due to the asymmetry of the Kullback-Leibler divergence, output provides both
\code{"KL_O_P"} which is equivalent to (Observed || Predicted) and
\code{"KL_P_O"} which is equivalent to (Predicted || Observed).

\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()}
function.

\code{group} optional grouping variable, if results should be displayed
by different groups. Has to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the
validation/holdout task (also includes the \code{none} alternative).
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of opts \code{choice} has to be column name of actual choice.

\code{epsilon} has to be a numeric input in case of 0 in the numerator or
denominator. 0 then will be replaced by \code{epsilon}. Default value
is \code{epsilon = 1e-5}, however, can be adopted.

\code{base} has to be a character string, deciding which logarithm base
you want to apply to calculate Kullback-Leibler. You can choose
between \eqn{log} and \eqn{log{_2}}. Default set to \eqn{log}.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# kl ungrouped - log
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log"
)

# kl ungrouped - log2
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log2"
)

# kl grouped - log + specifying epsilon
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log",
  group = Group,
  epsilon = 1e-8
)

# kl grouped - log2
kl(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  base = "log2",
  group = Group
)
}

}
\references{
{

Ding, Min, John R. Hauser, Songting Dong, Daria Dzyabura, Zhilin Yang,
SU Chenting, and Steven P. Gaskin. (2011).
Unstructured Direct Elicitation of Decision Rules. \emph{Journal of
Marketing Research 48}(1): 116-27. \verb{https://doi.org/10.1509/jmkr.48.1.116}.


Drost, Hajk-Georg. (2018). Philentropy: Information Theory and Distance
Quantification with R. \emph{Journal of Open Source Software 3}(26),
765, \verb{https://joss.theoj.org/papers/10.21105/joss.00765}.

}
}
