% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KL.R
\name{kl}
\alias{kl}
\title{Kullback-Leibler Divergence}
\usage{
kl(data, id, Group = NULL, opts, choice, epsilon = NULL)
}
\arguments{
\item{data}{data frame with all relevant variables.}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"kl"} by group.}

\item{opts}{vector of column indexes of the alternatives included in the
validation/holdout task.}

\item{choice}{vector of column index of the actual choice.}

\item{epsilon}{vector of noise that should be added to 0 values, per default set to 1e-05}
}
\value{
a data frame
}
\description{
Function to measure the Kullback-Leibler Divergence of a validation task
}
\details{
Kullback-Leibler-Divergence which measures the divergence between the actual choice distribution and the predicted
choice distribution (Ding et al., 2011; Drost, 2018). Currently only provides
the deviation measured based on \eqn{log{_2}} algorithm.

Due to the asymmetry of the Kullback-Leibler divergence, output provides both
\code{"KL_O_P"} which is equivalent to (Observed || Predicted) and
\code{"KL_P_O"} which is equivalent to (Predicted || Observed).

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{id} needs to be the column index of the id (unique for each participant)
in \code{data}.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a vector of the column index of \code{Group}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).

\code{choice} specifies the column index of the actual choice.
Input of opts \code{choice} needs to be the column index of actual choice.

\code{epsilon} needs to be a numeric input in case of 0 in the numerator or denominator. 0
then will be replaced by \code{epsilon}. Default value is \code{epsilon = .1e-5}, however, can
be adopted.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

kl(data = HOT, id = 1, opts = c(2:9), choice = 10)
}

\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

kl(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)
}

}
\references{
{

Ding, Min, John R. Hauser, Songting Dong, Daria Dzyabura, Zhilin Yang, SU Chenting, and Steven P. Gaskin. (2011).
Unstructured Direct Elicitation of Decision Rules. \emph{Journal of Marketing Research 48}(1): 116-27. \verb{https://doi.org/10.1509/jmkr.48.1.116}.

Drost, Hajk-Georg. (2018). Philentropy: Information Theory and Distance Quantification with R. \emph{Journal of Open Source Software 3}(26), 765, \verb{https://joss.theoj.org/papers/10.21105/joss.00765}.


}
}
