% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{Function to calculate mean absolute error of validation task}
\usage{
mae(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to calculate \code{mae()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}
}
\value{
a tibble
}
\description{
\code{mae()} calculates the mean absolute error of a
validation task, which is the aggregated deviation between predicted and
stated share of alternatives in the validation task.
}
\details{
Mean absolute error calculates the deviation between predicted and
stated (i.e., actual) choice share. It is an aggregated value across all
alternatives in the validation task.

\code{data} a \code{data.frame} object including the alternatives shown in the
validation task. It can be created using the \code{create_hot()} function.

\code{group} optional grouping variable(s) to display results by group(s).
different groups. Has to be the column name(s) of variables in \code{data}.

\code{opts} to specify the different alternatives in the
validation task (also includes the \code{none} alternative).

\code{choice} to specify the column of actual choice in the validation
task. The input of \code{choice} has to be the column name of actual choice.
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  choice = "hot",
  varskeep = "group"
)

# mae - without group argument defined
mae(
  data = hot,
  opts = c(option_1:none),
  choice = choice
)

# mae - with group argument defined
mae(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  group = group
)

}
