% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mae.R
\name{mae}
\alias{mae}
\title{Function to calculate mean absolute error of validation task}
\usage{
mae(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{mae()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}
}
\value{
a tibble
}
\description{
\code{mae()} measures the mean absolute error of a
validation task, i.e., aggregated deviation between predicted and
stated share of alternatives in the validation task.
}
\details{
Mean absolute error (MAE) calculates the deviation between predicted and
stated (actual) choice share. It is an aggregated value across all
alternatives in the validation task.

\code{data} a data frame including the alternatives shown in the validation task.
Can be created using the \code{create_hot()} function.

\code{group} optional grouping variable(s), if results should be displayed by
different groups. Has to be column name(s) of variables in \code{data}.

\code{opts} to specify the different alternatives in the validation task.
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice in the validation
task. Input of \code{choice} has to be column name of actual choice.
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  choice = "hot",
  varskeep = "group"
)

# mae - without group argument defined
mae(
  data = hot,
  opts = c(option_1:none),
  choice = choice
)

# mae - with group argument defined
mae(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  group = group
)

}
