% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{a data frame}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"mae"} by group.}

\item{opts}{vector of column indexes of the alternatives included in the
validation/holdout task.}

\item{choice}{vector of column index of the actual choice.}
}
\value{
a data frame
}
\description{
Function to measure the mean absolute error of a validation/ holdout task.
Calculates the averaged absolute error, i.e., deviation between predicted and
stated share of alternatives in the validation/ holdout task.
}
\examples{
library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

mae(data = HOT, id = 1, opts = c(2:9), choice = 10)

library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

mae(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
