% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{a data frame}

\item{id}{column index of the \code{id} variable}

\item{Group}{optional grouping variable to get hit rate by group}

\item{opts}{column indexes of the options included in the holdout task}

\item{choice}{column index of the actual choice}
}
\value{
a data frame
}
\description{
Function to measure the mean absolute error of a holdout task
}
\examples{
library(ValiDatHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

mae(data = HOT, id = 1, opts = c(2:9), choice = 10)

library(ValiDatHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

mae(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
