% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MAE.R
\name{mae}
\alias{mae}
\title{Mean Absolute Error}
\usage{
mae(data, group, opts, choice)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)
to get \code{"mae"} by group(s)}

\item{opts}{column names of the alternatives included in the
validation/holdout task}

\item{choice}{column name of the actual choice}
}
\value{
a tibble
}
\description{
Function to measure the mean absolute error of a validation/holdout task.
Calculates the averaged absolute error, i.e., deviation between predicted and
stated share of alternatives in the validation/holdout task.
}
\details{
Mean absolute error (MAE) calculates the deviation between predicted and
stated (actual) choice share. It is an aggregated value across all alternatives
in the validation/holdout task.

\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{group} optional grouping variable, if results should be displayed by different groups.
Has to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task.
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of opts \code{choice} has to be column name of actual choice.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# mae ungrouped
mae(data = HOT, opts = c(Option_1:None), choice = choice)

# mae grouped
mae(data = HOT, opts = c(Option_1:None), choice = choice, group = Group)
}
}
