% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MedAE.R
\name{MedAE}
\alias{MedAE}
\title{Median Absolute Error}
\usage{
MedAE(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{a data frame}

\item{id}{column index of the \code{id} variable}

\item{Group}{optional grouping variable to get hit rate by group}

\item{opts}{column indexes of the options included in the holdout task}

\item{choice}{column index of the actual choice}
}
\value{
xyz
}
\description{
Function to measure the Median absolute error of a holdout task
}
