% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mhp.R
\name{mhp}
\alias{mhp}
\title{Function to calculate Mean Hit Probability (MHP) of validation task}
\usage{
mhp(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{mhp()} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation task.}

\item{choice}{Column name of the actual choice in the validation
task.}
}
\value{
a tibble
}
\description{
\code{mhp()} measures the mean hit probability of a validation
task.
}
\details{
Mean hit probability (MHP) measures the averaged hit probability of
participants actual choices in the validation task.

\code{data} a data.frame object including the alternatives shown in the
validation task. Can be created using the \code{create_hot()} function.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{opts} is required to specify the different alternatives in the validation
task. Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice in the validation
task. Input of opts \code{choice} has to be column name of actual choice.

Output will display both mean hit probability and its corresponding standard
error (both in percentage).
}
\examples{

hot <- create_hot(
  data = maxdiff,
  id = "id",
  none = "none",
  prod.levels = list(2, 9, 10, 14, 15, 16, 17),
  method = "maxdiff",
  choice = "hot",
  varskeep = "group"
)

# mhp - without group argument defined
mhp(
  data = hot,
  opts = c(option_1:none),
  choice = choice
)

# mhp - with group argument defined
mhp(
  data = hot,
  opts = c(option_1:none),
  choice = choice,
  group = group
)

}
