% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHP.R
\name{mhp}
\alias{mhp}
\title{MHP}
\usage{
mhp(data, group, opts, choice)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)
to get \code{"mhp"} by group(s)}

\item{opts}{column names of the alternatives included in the
validation/holdout task}

\item{choice}{column name of the actual choice}
}
\value{
a tibble
}
\description{
Function to measure the mean hit probability of a validation/
holdout task. Calculates the hit/ choice probability of the actual choice
in the validation/ holdout task.
}
\details{
Mean hit probability (MHP) measures the averaged hit probability of participants actual
choices in the validation/ holdout task.

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{group} optional Grouping variable, if results should be display by different groups.
Needs to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the \code{none} alternative).
Input of \code{opts} needs to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of opts \code{choice} needs to be column name of actual choice.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20, varskeep = 21
)

# mhp ungrouped
mhp(data = HOT, opts = c(Option_1:None), choice = choice)

# mhp grouped
mhp(data = HOT, opts = c(Option_1:None), choice = choice, group = Group)

}


}
