% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHP.R
\name{mhp}
\alias{mhp}
\title{MHP}
\usage{
mhp(data, group, opts, choice)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s)
to get \code{mhp} by group(s).}

\item{opts}{Column names of the alternatives included in the
validation/holdout task.}

\item{choice}{Column name of the actual choice.}
}
\value{
a tibble
}
\description{
\code{mhp} measures the mean hit probability of a validation/holdout
task. Calculates the hit/choice probability of the actual choice
in the validation/holdout task.
}
\details{
Mean hit probability (MHP) measures the averaged hit probability of
participants actual choices in the validation/holdout task.

\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()}
function.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the
validation/holdout task.
Input of \code{opts} has to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of opts \code{choice} has to be column name of actual choice.

Output will display both mean hit probability and its corresponding standard
error.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  none = 19,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# mhp ungrouped
mhp(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice
)

# mhp grouped
mhp(
  data = HOT,
  opts = c(Option_1:None),
  choice = choice,
  group = Group
)
}

}
