% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MHP.R
\name{MHP}
\alias{MHP}
\title{MHP}
\usage{
MHP(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{a data frame}

\item{id}{column index of the \code{id} variable}

\item{Group}{optional grouping variable to get hit rate by group}

\item{opts}{column indexes of the options included in the holdout task}

\item{choice}{column index of the actual choice}
}
\value{
xyz
}
\description{
Function to measure the mean hit probability of a holdout task
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff")
MHP(data = HOT, id = 1, opts = c(2:9), choice = 10)


library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19, id = 1,
          prod = 7, x = list(3, 10, 11, 15, 16, 17, 18),
          choice = 20, method = "MaxDiff", varskeep = 21)
MHP(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
