% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_scores.R
\name{prob_scores}
\alias{prob_scores}
\title{Function to calculate probability scores for (anchored) MaxDiff}
\usage{
prob_scores(
  data,
  group = NULL,
  items,
  set.size,
  res = c("agg", "ind"),
  anchor = NULL
)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s).}

\item{items}{A vector with the column names of the items.}

\item{set.size}{A vector that specifies the size of the choice set.}

\item{res}{A vector indicating whether individual results (\code{ind}) or
aggregated (\code{agg}) results should be returned.}

\item{anchor}{An optional variable to specify anchor variable.}
}
\value{
a tibble
}
\description{
Function to calculate probability scores for (anchored) MaxDiff
}
\details{
\code{prob_scores()} converts raw utilities from a MaxDiff to probability scores.
Probability scores for the unanchored MaxDiff are calculated according to the
formula provided by Chrzan & Orme (2019, p. 56):
\eqn{\frac{e^U}{(e^U + (a - 1)}}, where \code{U} is the raw utility of the
item and \code{a} is the number of items shown per choice task.

For anchored MaxDiff the following formula is applied
\eqn{\frac{e^U}{(e^U + (a - 1)} * 100 / (1 / a)} (Chrzan & Orme, 2019,
pp. 59-60).

\code{data} a \code{data.frame} object with the items (raw utilities).

\code{group} optional grouping variable(s) to display results by group(s).
Has to be the column name(s) of variables in \code{data}.

\code{items} specifies the items of the MaxDiff. Has to be column name of
variables in \code{data}.

\code{set.size} specifies the size of the choice sets (i.e., MaxDiff tasks, how
many items were shown in one task). Input must be an integer.

\code{res} specifies whether results should be aggregated across all participants
or across \code{group} (\code{res} needs to be set to \code{agg}) or if scores should be
converted for individuals only (\code{res} needs to be set to \code{ind}).

\code{anchor} only required if anchored MaxDiff is applied.
Input for \code{anchor} has to be the column name of variables in \code{data}.
}
\examples{

# probability scores for unanchored maxdiff - without group argument defined
prob_scores(
  data = maxdiff,
  items = c(option_01:option_16),
  set.size = 4,
  res = "agg"
)

# probability scores for anchored maxdiff - without group argument defined
prob_scores(
  data = maxdiff,
  items = c(option_01:none),
  set.size = 4,
  anchor = none,
  res = "agg"
)

}
\references{
{

Chrzan, K., & Orme, B. K. (2019). \emph{Applied MaxDiff: A Practitioner’s
Guide to Best-Worst Scaling} Provo, UT: Sawtooth Software.

}
}
\seealso{
{
\code{\link[=zero_anchored]{zero_anchored()}} for zero-anchored interval
scores for MaxDiff
\code{\link[=att_imp]{att_imp()}} for attribute importance scores for (A)CBC
\code{\link[=zc_diffs]{zc_diffs()}} for zero-centered diff scores for (A)CBC
}
}
