% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_scores.R
\name{prob_scores}
\alias{prob_scores}
\title{Title}
\usage{
prob_scores(data, group = NULL, items, set.size, anchor = NULL)
}
\arguments{
\item{data}{a}

\item{group}{v}

\item{items}{v}

\item{set.size}{c}

\item{anchor}{c}
}
\value{
a tibble
}
\description{
Title
}
\examples{
\dontrun{
prob_scores(data = MaxDiff,
            items = c(Option_01:Option_16),
            set.size = 4,
            anchor = "none"
)
}
}
