% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_scores.R
\name{prob_scores}
\alias{prob_scores}
\title{Probability Scores for (anchored) MaxDiff}
\usage{
prob_scores(data, group = NULL, items, set.size, anchor = NULL)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)}

\item{items}{specifies the items}

\item{set.size}{size of the choice sets}

\item{anchor}{optional variable to specify anchor variable}
}
\value{
a tibble
}
\description{
Probability Scores for (anchored) MaxDiff
}
\examples{
\dontrun{

prob_scores(data = MaxDiff,
            items = c(Option_01:Option_16),
            set.size = 4,
            anchor = "none"
)

prob_scores(data = MaxDiff,
            items = c(Option_01:Option_16),
            set.size = 4,
            anchor = "none"
)
}
}
