% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_scores.R
\name{prob_scores}
\alias{prob_scores}
\title{Function to calculate probability scores for (anchored) MaxDiff}
\usage{
prob_scores(
  data,
  group = NULL,
  items,
  set.size,
  res = c("agg", "ind"),
  anchor = NULL
)
}
\arguments{
\item{data}{A data frame with all relevant variables.}

\item{group}{Optional column name(s) to specify grouping variable(s).}

\item{items}{Vector that specifies the items.}

\item{set.size}{A vector that specifies size of the choice set.}

\item{res}{A vector indicating whether individual shares (\code{ind}) or
aggregated (\code{agg}) shares should be returned.}

\item{anchor}{An optional variable to specify anchor variable.}
}
\value{
a tibble
}
\description{
Function to calculate probability scores for (anchored) MaxDiff
}
\details{
\code{prob_scores} converts raw utilities of a MaxDiff to probability scores.
Probability scores for the unanchored MaxDiff are calculated according to the formula provided by
Chrzan & Orme (2019, p. 56): \eqn{\frac{e^U}{(e^U + (a - 1)}}, where \emph{U} is the
raw utility of the item and \emph{a} is the number of items shown per choice task.

For anchored MaxDiff the following formula is applied \eqn{\frac{e^U}{(e^U + (a - 1)} * 100 / (1 / a)}
(Chrzan & Orme, 2019, pp. 59-60).

\code{data} has to be a data frame with the attributes. Items have
to be the raw utilities.

\code{group} optional grouping variable, if results should be displayed by
different groups. Has to be column name of variables in \code{data}.

\code{items} specifies the items of the MaxDiff.
Input for \code{items} has to be variable names.

\code{set.size} specifies the size of the choice sets (how many items were
shown in one task). Input needs to be a whole number.

\code{res} specifies whether results should be aggregated across all participants
or across \code{group} (\code{res} needs to be set to \code{agg}) or if it scores
should be converted for individuals only

\code{anchor} only needs to be specified if anchored MaxDiff is applied. Input
}
\examples{
\dontrun{

# probability scores for unanchored MaxDiff - without Group defined
prob_scores(
  data = MaxDiff,
  items = c(Option_01:Option_16),
  set.size = 4,
  res = "agg"
)

# probability scores for unanchored MaxDiff - with Group defined
prob_scores(
  data = MaxDiff,
  group = Group,
  items = c(Option_01:Option_16),
  set.size = 4,
  res = "agg"
)


# probability scores for anchored MaxDiff - without Group defined
prob_scores(
  data = MaxDiff,
  items = c(Option_01:none),
  set.size = 4,
  anchor = none,
  res = "agg"
)

# probability scores for anchored MaxDiff - with Group defined
prob_scores(
  data = MaxDiff,
  group = Group,
  items = c(Option_01:none),
  set.size = 4,
  anchor = none,
  res = "agg"
)
}

}
\references{
{

Chrzan, K., & Orme, B. K. (2019). \emph{Applied MaxDiff: A Practitioner’s
Guide to Best-Worst Scaling} Provo, UT: Sawtooth Software.

}
}
\seealso{
{
\code{\link[=zero_anchored]{zero_anchored}} for zero-anchored interval scores for MaxDiff
\code{\link[=att_imp]{att_imp}} for attribute importance scores for (A)CBC
\code{\link[=zc_diffs]{zc_diffs}} for zero-center diff scores for (A)CBC
}
}
