% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach.R
\name{reach}
\alias{reach}
\title{Percentage of participants that are reached by assortment}
\usage{
reach(data, id, Group = NULL, None, method, bundles)
}
\arguments{
\item{data}{data frame including Holdout Options and actual \code{"choice"} and \code{"Group"} if optional argument is defined}

\item{id}{column index}

\item{Group}{optional grouping variable to get accuracy by group}

\item{None}{column index}

\item{method}{which method should be used, can be either \code{"First choice"} (only the bundle with highest utility is marked as purchase option if its above \code{"None"} utility) or \code{"threshold"} (all bundles with utility above \code{"None"} utility are marked as purchase option)}

\item{bundles}{column indexes of the bundles included that should be included in the assortment}
}
\value{
a data frame
}
\description{
test
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19,
         id = 1, prod = 7,
         prod.levels = list(3, 10, 11, 15, 16, 17, 18),
         choice = 20, method = "MaxDiff"
)
reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "threshold")
reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "First Choice")

library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19,
         id = 1, prod = 7,
         prod.levels = list(3, 10, 11, 15, 16, 17, 18),
         choice = 20, method = "MaxDiff", varskeep = 21
)
reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "threshold", Group = 10)
reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "First Choice", Group = 10)

}
