% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach.R
\name{reach}
\alias{reach}
\title{Percentage of participants that are reached by assortment}
\usage{
reach(data, group, none, opts)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)
to get \code{"reach"} by group(s)}

\item{none}{column name of none alternative}

\item{opts}{column names of the alternatives included in the assortment}
}
\value{
a tibble
}
\description{
Reach function of TURF analysis to measure the number of  the averaged percentage of how many participants you can reach
(at least one of the products resemble a purchase option) is reached with a specific product bundle assortment.
}
\details{
\code{"reach"} calculates the the percentage of consumers that would be reached with the
product assortment you are testing. The current logic of \code{reach()}
is that it needs to exceed a threshold. In the case of \code{reach()}
this threshold is referred to the \code{none} argument in \code{data}.


\code{data} needs to be a data frame including the alternatives that should be tested

\code{group} optional Grouping variable, if results should be display by different conditions.
Needs to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the
product assortment that should be considered.
Input of \code{opts} needs to be column names of variables in \code{data}.

\code{none} to specify column name of the \code{none} alternative in the
validation/holdout task.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20, varskeep = 21
)

# reach ungrouped
reach(data = HOT, opts = c(Option_1, Option_2, Option_6), none = None)

# reach grouped
reach(data = HOT, opts = c(Option_1, Option_2, Option_6), none = None, group = Group)

}

}
