% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach.R
\name{Reach}
\alias{Reach}
\title{Percentage of participants that are reached by assortment}
\usage{
Reach(data, id, Group = NULL, None, method, bundles)
}
\arguments{
\item{data}{data frame including Holdout Options and actual \code{"choice"} and \code{"Group"} if optional argument is defined}

\item{id}{column index}

\item{Group}{optional grouping variable to get accuracy by group}

\item{None}{column index}

\item{method}{which method should be used, can be either \code{"First choice"} (only the bundle with highest utility is marked as purchase option if its above \code{"None"} utility) or \code{"threshold"} (all bundles with utility above \code{"None"} utility are marked as purchase option)}

\item{bundles}{column indexes of the bundles included that should be included in the assortment}
}
\value{
xyz
}
\description{
test
}
