% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Reach.R
\name{reach}
\alias{reach}
\title{Percentage of participants that are reached by assortment}
\usage{
reach(data, id, Group = NULL, None, method, bundles)
}
\arguments{
\item{data}{data data frame with all relevant variables.}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"reach"} by group.}

\item{None}{vector of column index of None alternative.}

\item{method}{character variable that needs to be specified to either \code{"First choice"} or \code{"threshold"}, please see Details.}

\item{bundles}{vector of column indexes of the bundles that should be included in the assortment.}
}
\value{
a data frame
}
\description{
Reach function of TURF analysis to measure the number of  the averaged percentage of how many participants you can reach
(at least one of the products resemble a purchase option) is reached with a specific product bundle assortment. Can either be calculated
using 'First Choice' or 'threshold' rule. See Details section for more
information and specifying the data in the correct way.
}
\examples{
library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "threshold")
reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "First Choice")

library(validateHOT)
data("MaxDiff")
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "threshold", Group = 10)
reach(data = HOT, id = 1, bundles = c(2, 3, 7), None = 9, method = "First Choice", Group = 10)

}
