% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Recall.R
\name{recall}
\alias{recall}
\title{Recall}
\usage{
recall(data, id, Group = NULL, opts, choice, None)
}
\arguments{
\item{data}{data frame including Holdout Options and actual \code{"choice"} and \code{"Group"} if optional argument is defined}

\item{id}{column index}

\item{Group}{optional grouping variable to get accuracy by group}

\item{opts}{column indexes of the options included in the holdout task}

\item{choice}{column index of the actual choice}

\item{None}{column index}
}
\value{
a data frame
}
\description{
recall is one of the 5 metrics of the confusion matrix
and is defined as \eqn{\frac{TP}{TP + FN}}, where TP =
True Positives, FN = False Negatives (see, e.g., Burger, 2018).
}
\details{
The current logic of \code{"recall"} is to provide whether a binary coded event is correctly predicted.
To use the function a \code{"None"} alternative needs to be in the script.
One potential usage is, for example, whether a buy or a no-buy condition
was predicted correctly. For example, you have three alternatives plus
a \code{"None"} alternative and you want to check whether a buy or no-buy was
correctly predicted. This function can be helpful when you test whether or
not your model significantly overestimates or underestimates, for example, a purchase likelihood.

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{id} needs to be the column index of the id (unique for each participant)
in \code{data}.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a vector of the column index of \code{Group}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).

\code{choice} specifies the column index of the actual choice.
Input of opts \code{choice} needs to be the column index of actual choice.

\code{None} specifies the column index of the \code{None} alternative in the
validation/holdout task. Needs to be specified for \code{recall()}.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

recall(data = HOT, id = 1, opts = c(2:9), choice = 10, None = 9)
}

\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

recall(data = HOT, id = 1, Group = 10, opts = c(2:9), choice = 11, None = 9)
}

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R: Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=accuracy]{accuracy}}
\code{\link[=f1]{f1}}
\code{\link[=precision]{precision}}
\code{\link[=specificity]{specificity}}
}
}
