% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMSE.R
\name{rmse}
\alias{rmse}
\title{Root Mean Squared Error}
\usage{
rmse(data, id, Group = NULL, opts, choice)
}
\arguments{
\item{data}{a data frame}

\item{id}{vector of column index of unique identifier in \code{data}.}

\item{Group}{optional vector of column number to specify grouping variable
to get \code{"rmse"} by group.}

\item{opts}{vector of column indexes of the alternatives included in the
validation/ holdout task.}

\item{choice}{vector of column index of the actual choice.}
}
\value{
a data frame
}
\description{
Function to measure the root mean squared error
 of a validation/ holdout task. Calculates the averaged
 root mean squared error of the stated and predicted share of alternatives
 in the validation/ holdout task.
}
\details{
Root mean squared error (RMSE) calculates the root mean squared error when comparing
the share of the actual choice in the holdout task and the predicted share.


\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{id} needs to be the column index of the id (unique for each participant)
in \code{data}.

\code{Group} optional Grouping variable, if results should be display by different conditions.
Input of \code{Group} needs to be a vector of the column index of \code{Group}.

\code{opts} is needed to specify the different alternatives in the validation/ holdout
task (also includes the None option).
Input of \code{opts} needs to be a vector with column index(es).

\code{choice} specifies the column index of the actual choice.
Input of opts \code{choice} needs to be the column index of actual choice.
}
\examples{
library(validateHOT)
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20
)

rmse(data = HOT, id = 1, opts = c(2:9), choice = 10)

library(validateHOT)
createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  varskeep = 21,
  choice = 20
)

rmse(data = HOT, id = 1, opts = c(2:9), choice = 11, Group = 10)

}
