% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShareOfPref.R
\name{shareofpref}
\alias{shareofpref}
\title{Share of Preferences of Options included in HOT}
\usage{
shareofpref(data, id, Group = NULL, opts)
}
\arguments{
\item{data}{a data frame}

\item{id}{column index of the \code{id} variable}

\item{Group}{optional grouping variable to get hit rate by group}

\item{opts}{column indexes of the options included in the holdout task}
}
\value{
a data frame or list
}
\description{
Function to measure the share of preferences of each option in the validation task
}
\examples{
library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19,
         id = 1, prod = 7,
         prod.levels = list(3, 10, 11, 15, 16, 17, 18),
         choice = 20, method = "MaxDiff"
)
shareofpref(data = HOT, id = 1, opts = c(2:9))

library(ValiDatHOT)
data(MaxDiff)
createHOT(data = MaxDiff, None = 19,
         id = 1, prod = 7,
         prod.levels = list(3, 10, 11, 15, 16, 17, 18),
         choice = 20, method = "MaxDiff", varskeep = 21
)
shareofpref(data = HOT, id = 1, opts = c(2:9), Group = 10)

}
