% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShareOfPref.R
\name{shareofpref}
\alias{shareofpref}
\title{Share of Preferences of Options included in HOT}
\usage{
shareofpref(data, group, opts)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)
to get \code{"shareofpref"} by group(s)}

\item{opts}{column names of the alternatives included in the
validation/holdout task}
}
\value{
a tibble
}
\description{
Function to measure the share of preferences of each option in a validation/holdout task
}
\details{
Share of Preference provides the aggregated share of each alternative in the
validation/holdout task as well as the lower and upper confidence interval
of each alternative which is calculated according to the following formula
\eqn{mean +/- 1.96 x \frac{sd}{\sqrt(n)}} (Orme, 2020, p. 94).

\code{data} has to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{group} optional grouping variable, if results should be displyed by different groups.
Has to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the simulation
task.
Input of \code{opts} has to be column names of variables in \code{data}.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20,
  varskeep = 21
)

# shareofpref ungrouped
shareofpref(data = HOT, opts = c(Option_1:None))

# shareofpref grouped
shareofpref(data = HOT, group = Group, opts = c(Option_1:None))
}

}
\references{
{

Orme, B. K. (2020). \emph{Getting Started with Conjoint Analysis:
Strategies for Product Design and Pricing Research}. 4th edition.
Manhattan Beach, CA: Research Publishers LLC.

}
}
