% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Specificity.R
\name{specificity}
\alias{specificity}
\title{Specificity}
\usage{
specificity(data, group, opts, choice, none)
}
\arguments{
\item{data}{data frame with all relevant variables}

\item{group}{optional column name(s) to specify grouping variable(s)
to get \code{"recall"} by group(s)}

\item{opts}{column names of the alternatives included in the
validation/holdout task}

\item{choice}{column name of the actual choice}

\item{none}{column name of none alternative}
}
\value{
a tibble
}
\description{
specificity is one of the 5 metrics of the confusion matrix
and is defined as \eqn{\frac{TN}{TN + FP}}, where TN =
True Negatives, FP = False Positives (see, e.g., Burger, 2018).
}
\details{
The current logic of \code{"recall"} is to provide whether a binary coded event is correctly predicted.
To use the function a \code{"none"} alternative needs to be in the script.
One potential usage is, for example, whether a buy or a no-buy condition
was predicted correctly. For example, you have three alternatives plus
a \code{"none"} alternative and you want to check whether a buy or no-buy was
correctly predicted. This function can be helpful when you test whether or
not your model significantly overestimates or underestimates, for example, a purchase likelihood.
This function was programmed to predict demand for a MaxDiff, CBC, or ACBC.

\code{data} needs to be a data frame including the alternatives shown in
the validation/holdout task. Can be created using the \code{createHOT()} function.

\code{group} optional Grouping variable, if results should be display by different conditions.
Needs to be column name of variables in \code{data}.

\code{opts} is needed to specify the different alternatives in the validation/holdout
task (also includes the \code{none} alternative).
Input of \code{opts} needs to be column names of variables in \code{data}.

\code{choice} to specify column of actual choice.
Input of opts \code{choice} needs to be column name of actual choice.

\code{none} to specify column name of the \code{none} alternative in the
validation/holdout task.
}
\examples{
\dontrun{
HOT <- createHOT(
  data = MaxDiff,
  id = 1,
  None = 19,
  prod = 7,
  prod.levels = list(3, 10, 11, 15, 16, 17, 18),
  method = "MaxDiff",
  choice = 20, varskeep = 21
)

# specificity ungrouped
specificity(data = HOT, opts = c(Option_1:None), choice = choice, none = None)

# specificity by group
specificity(data = HOT, opts = c(Option_1:None), choice = choice, none = None, group = Group)
}

}
\references{
{

Burger, S. V. (2018). \emph{Introduction to Machine Learning with R: Rigorous Mathematical Analysis}. O'Reilly.

}
}
\seealso{
{
\code{\link[=accuracy]{accuracy}}
\code{\link[=f1]{f1}}
\code{\link[=precision]{precision}}
\code{\link[=recall]{recall}}
}
}
