% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turf.R
\name{turf}
\alias{turf}
\title{Function to run Total Unduplicated Reach and Frequency analysis}
\usage{
turf(
  data,
  opts,
  none,
  size,
  fixed = NULL,
  prohib = NULL,
  approach = c("thres", "fc")
)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{opts}{Column names of the alternatives included in the assortment.}

\item{none}{Column name of the none / threshold alternative.}

\item{size}{An integer to determine the size of the assortment.}

\item{fixed}{An optional vector to determine alternatives / items that have
to be included in the assortment.}

\item{prohib}{An optional list containing vectors to determine
prohibitions, i.e., alternatives / items that are not allowed to be together
in one assortment.}

\item{approach}{A character whether to run the first-choice approach
(\code{'fc'}) or threshold approach (\code{'thres'}).}
}
\value{
data frame
}
\description{
T(otal) U(nduplicated) R(each) and F(requency) is a "product line extension
model" (Miaoulis et al., 1990, p. 29). For every possible combination,
\code{turf()} calculates the reach and frequency.
Participants are reached, if at least one of the alternatives in a
combination has a higher utility than \code{none}. On the contrary, frequency
calculates the average number of alternatives that have a
higher utility than \code{none}.
}
\details{
\code{data} a \code{data.frame} object containing the alternatives that
should be tested.

\code{opts} to specify the different alternatives in the product assortment that
should be considered. Input of \code{opts} has to be the column names of variables
in \code{data}.

\code{none} to specify the column name of the \code{none} alternative in \code{data} (i.e.,
variable that needs to be exceeded).

\code{size} an integer to determine the size of the assortment.

\code{fixed} has to be a vector of variables that are fixed in the
assortment, i.e., they have to be part of the assortment.

\code{prohib} has to be a list of vectors of variables that are prohibited
in the assortment, i.e., alternatives that are not allowed to be together in
one assortment.

\code{approach} character defining whether first choice (\code{approach = 'fc'})
or threshold approach (\code{approach = 'thres'}) should be applied for running
\code{turf()}. If \code{approach = 'fc'}, participants are considered to be
reached, if their alternative with the highest utility is included in the
assortment and this alternative's utility is larger than the threshold's
utility (Chrzan & Orme, 2019, p. 111).
On the contrary, if \code{approach = 'thres'}, participants are considered
to be reached, if the utility of one product is higher than the one of the
\code{none} alternative (Chrzan & Orme, 2019, p. 112).
If \code{approach = 'fc'}, \code{reach} equals \code{freq} since participants have at
maximum their most preferred alternative that exceeds the \code{none}
alternative.
}
\examples{


# turf no fixed alternatives + no prohibitions
t1 <- turf(
  data = maxdiff,
  opts = c(option_01:option_16),
  none = none,
  size = 3L,
  approach = "thres"
)

head(t1)

# turf alternative 4 and 5 fixed, no prohibitions
t2 <- turf(
  data = maxdiff,
  opts = c(option_01:option_16),
  none = none,
  size = 4L,
  fixed = c("option_04", "option_05"),
  approach = "thres"
)

head(t2)

#' # turf alternative 4 and 5 fixed, 2 and 9 not allowed together
t3 <- turf(
  data = maxdiff,
  opts = c(option_01:option_16),
  none = none,
  size = 4L,
  fixed = c("option_04", "option_05"),
  prohib = list(c("option_02", "option_09")),
  approach = "thres"
)

head(t3)

}
\references{
{

Chrzan, K., & Orme, B. K. (2019). \emph{Applied MaxDiff: A Practitioner’s
Guide to Best-Worst Scaling} Provo, UT: Sawtooth Software.

Miaoulis, G., Parsons, H., & Free, V. (1990). Turf: A New Planning Approach
for Product Line Extensions. \emph{Marketing Research 2} (1): 28-40.

}
}
\seealso{
{
\code{\link[=turf_ladder]{turf_ladder()}}
}
}
