% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zc_diffs.R
\name{zc_diffs}
\alias{zc_diffs}
\title{Function to calculate zero-centered diffs for (A)CBCs}
\usage{
zc_diffs(
  data,
  group = NULL,
  attrib,
  coding,
  interpolate.levels = NULL,
  res = c("agg", "ind"),
  none = NULL
)
}
\arguments{
\item{data}{A data frame containing all relevant variables.}

\item{group}{Optional variable to to get \code{zc_diffs()} by group(s).}

\item{attrib}{A list that specifies the attribute levels for each attribute.}

\item{coding}{A vector of the coding of each attribute, '0' = part-worth
coding, '1' = linear coding, or '2' = piecewise coding.}

\item{interpolate.levels}{A list of the levels of the attribute that should
be interpolated. These levels must match those specified in model
estimation (e.g., if you scale or center attribute levels before estimation,
insert the scaled or centered levels). Ensure to provide the entire list.
It has to be specified only for the variables that are coded as '1' (linear).}

\item{res}{A character vector indicating whether individual zero-centered
diffs  (\code{ind}) or aggregated (\code{agg}) zero-centered diffs should be returned.}

\item{none}{An optional vector to specify the \code{none}
alternative in \code{data}.}
}
\value{
a tibble
}
\description{
Function to calculate zero-centered diffs for (A)CBCs
}
\details{
\code{zc_diffs()} converts raw utilities of a CBC or an ACBC to
zero-centered diffs (see, e.g., Orme, 2020, p. 78).

\code{data} a \code{data.frame} object containing all the relevant attribute levels.
Attribute levels have to be the raw utilities.

\code{group} optional grouping variable(s) to display results by group(s).
Has to be the column name(s) of variables in \code{data}.

\code{attrib} specifies the attribute levels for each alternative.
Input for \code{attrib} has to be a list.

\code{coding} indicates the attribute coding. \code{0} to indicate part-worth coding,
\code{1} for linear coding, or \code{2} for piecewise coding.

\code{interpolate.levels} is required for linear-coded variables.
If scaled or centered values were used for Hierarchical Bayes
estimation, these have to be specified in this case.
All values have to be specified. For example, if one linear-coded attribute
has 5 levels, all 5 levels have to be inserted.

\code{res} specifies whether results should be aggregated across all participants
or across \code{group} (\code{res} needs to be set to \code{agg}) or if scores
should be converted for individuals only (\code{ind}).

\code{none} specifies whether the none option was included. If yes,
column name or column index of the \code{none} alternative needs to be specified.
Otherwise leave it empty.
}
\examples{

# zero-centered scores for CBC (only part-worth) - without group
# argument defined
zc_diffs(
  data = cbc,
  attrib = list(
    paste0("att1_lev", c(1:3)),
    paste0("att2_lev", c(1:2)),
    paste0("att3_lev", c(1:4)),
    paste0("att4_lev", c(1:4)),
    paste0("att5_lev", c(1:2)),
    paste0("att6_lev", c(1:4)),
    paste0("att7_lev", c(1:6)),
    paste0("att8_lev", c(1:6)),
    paste0("price_", c(1:6))
  ),
  coding = c(rep(0, times = 9)),
  none = "none",
  res = "agg"
)

\dontrun{

# zero-centered scores for CBC (incl. linear coded attribute) - without
# group argument defined
zc_diffs(
  data = cbc_linear,
  attrib = list(
    paste0("att1_lev", c(1:3)),
    paste0("att2_lev", c(1:2)),
    paste0("att3_lev", c(1:4)),
    paste0("att4_lev", c(1:4)),
    paste0("att5_lev", c(1:2)),
    paste0("att6_lev", c(1:4)),
    paste0("att7_lev", c(1:6)),
    paste0("att8_lev", c(1:6)),
    "price"
  ),
  coding = c(rep(0, times = 8), 1),
  interpolate.levels = list(c(seq(from = 175.99, to = 350.99, by = 35))),
  none = "none",
  res = "agg"
)
}

}
\references{
{

Orme, B. K. (2020). \emph{Getting Started with Conjoint Analysis: Strategies
for Product Design and Pricing Research}. 4th edition. Manhattan Beach, CA:
Research Publishers LLC.

}
}
\seealso{
{
\code{\link[=att_imp]{att_imp()}} for attribute importance scores for (A)CBC
\code{\link[=prob_scores]{prob_scores()}} for probability scores for MaxDiff
\code{\link[=zero_anchored]{zero_anchored()}} for zero-anchored interval
scores for MaxDiff
}
}
